### 2025-06-08 854300d
Update get_fairness_metrics.R
for better presentation
