+++ R/get_fairness_metrics.R
@@ -98,7 +98,6 @@ get_fairness_metrics <- function(data,
     alpha = alpha,
     digits = digits,
     message = FALSE
-
   )
 
   if (!(is.null(group2) & is.null(condition))) {
@@ -119,22 +118,6 @@ get_fairness_metrics <- function(data,
       digits = digits,
       message = FALSE
     )
-
-    split_condition <- strsplit(condition, split = '')[[1]]
-    if (split_condition[1] == ">" | split_condition[1] == "=") {
-      cond_stats_parity[["Full Metric Name"]] <- paste0("Conditional Statistical Parity (",
-                                              group2,
-                                              " ",
-                                              condition,
-                                              ")")
-    } else {
-      cond_stats_parity[["Full Metric Name"]] <- paste0("Conditional Statistical Parity (",
-                                              condition,
-                                              " ",
-                                              group2,
-                                              ")")
-    }
-
   }
   eq_opp <- eval_eq_opp(
     data = data,
@@ -245,56 +228,48 @@ get_fairness_metrics <- function(data,
     message = FALSE
   )
 
-  # Rename metrics so that its easy to read
-  stats_parity[["Full Metric Name"]] <- "Statistical Parity"
-  eq_opp[["Full Metric Name"]] <- "Equal Opportunity"
-  pred_equality[["Full Metric Name"]] <- "Predictive Equality"
-  pos_class_bal[["Full Metric Name"]] <- "Balance for Positive Class"
-  neg_class_bal[["Full Metric Name"]] <- "Balance for Negative Class"
-  pos_pred_parity[["Full Metric Name"]] <- "Positive Predictive Parity"
-  neg_pred_parity[["Full Metric Name"]] <- "Negative Predictive Parity"
-  bs_parity[["Full Metric Name"]] <- "Brier Score Parity"
-  acc_parity[["Full Metric Name"]] <- "Overall Accuracy Parity"
-  treatment_equality[["Full Metric Name"]] <- "Treatment Equality"
-
+  # Model performance summary
   if (!(is.null(group2) & is.null(condition))) {
-    result <- do.call(
-      rbind,
-      list(
-        stats_parity,
-        cond_stats_parity,
-        eq_opp,
-        pred_equality,
-        pos_class_bal,
-        neg_class_bal,
-        pos_pred_parity,
-        neg_pred_parity,
-        bs_parity,
-        acc_parity,
-        treatment_equality
-      )
-    )
-
+    summary <- rbind(stats_parity, cond_stats_parity,
+                     eq_opp, pred_equality, pos_class_bal,
+                     neg_class_bal, pos_pred_parity, neg_pred_parity, bs_parity,
+                     acc_parity, treatment_equality)
+    fairness_summary <- summary[, -c(1:3)]
+    fairness_summary <- cbind( "Metirc" = c("Statistical Parity",
+                                            "Conditional Statistical Parity",
+                                            "Equal Opportunity",
+                                            "Predictive Equality",
+                                            "Balance for Positive Class",
+                                            "Balance for Negative Class",
+                                            "Positive Predictive Parity",
+                                            "Negative Predictive Parity",
+                                            "Brier Score Parity",
+                                            "Overall Accuracy Parity",
+                                            "Treatment Equality"),
+                               fairness_summary)
   } else{
-    result <- do.call(
-      rbind,
-      list(
-        stats_parity,
-        eq_opp,
-        pred_equality,
-        pos_class_bal,
-        neg_class_bal,
-        pos_pred_parity,
-        neg_pred_parity,
-        bs_parity,
-        acc_parity,
-        treatment_equality
-      )
-    )
-
[…]
