### 2025-06-08 8c2d307
Update fairmetrics.Rmd
