+++ vignettes/fairmetrics.Rmd
@@ -180,7 +180,7 @@ test_data %>%
 To calculate various fairness metrics for the model, we pass our test data with its predicted results into the `get_fairness_metrics` function.
 
 ```{r}
-fairness_result <- fairmetrics::get_fairness_metrics(
+result <- fairmetrics::get_fairness_metrics(
   data = test_data,
   outcome = "day_28_flg",
   group = "gender",
@@ -190,7 +190,7 @@ fairness_result <- fairmetrics::get_fairness_metrics(
   cutoff = cut_off
  )
 
-kableExtra::kable(fairness_result, booktabs = TRUE, escape = FALSE) %>%
+kableExtra::kable(result$fairness, booktabs = TRUE, escape = FALSE) %>%
   kableExtra::kable_styling(full_width = FALSE) %>%
   kableExtra::pack_rows("Independence-based criteria", 1, 2) %>%
   kableExtra::pack_rows("Separation-based criteria", 3, 6) %>%
