### 2025-06-11 8047dd4
incrementing package version. Re-documenting
