+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: fairmetrics
 Title: Fairness Evaluation Metrics with Confidence Intervals
-Version: 1.0.2
+Version: 1.0.3
 Authors@R: 
     c(
     person("Jianhui", "Gao", email = "jianhui.gao@mail.utoronto.ca", role = c("aut"),comment = c(ORCID = "0000-0003-0915-1473")),
