+++ pkgdown.yml
@@ -3,4 +3,4 @@ pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   fairmetrics: fairmetrics.html
-last_built: 2025-06-11T00:23Z
+last_built: 2025-06-11T01:41Z
