### 2025-06-12 b2958e1
shortened paper
