+++ utils/joss_paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'fairmetrics: An R package for group fairness evaluation'
+title: '{fairmetrics}: An R package for group fairness evaluation'
 tags:
   - R
   - Fairness
@@ -18,105 +18,57 @@ authors:
 affiliations:
   - name: "Department of Statistical Science, University of Toronto"
     index: 1
-date: "2025-06-02"
+date: "2025-06-10"
 bibliography: paper.bib
+nocite: |
+  @*
 output:
-  rticles::joss_article
-  #rticles::arxiv_article
+  rticles::joss_article:
+    keep_tex: true
   # md_document:
-  #   preserve_yaml: TRUE
-  #   variant: "markdown_strict"
+  #    keep_tex: true
+  #    preserve_yaml: TRUE
+  #    variant: "markdown_strict"
 journal: JOSS
 ---
 
 # Summary
 
-Fairness is a growing area of machine learning (ML) that focuses on ensuring models do not produce systematically biased outcomes for specific groups, particularly those defined by protected attributes such as race, gender, or age. Evaluating fairness is a critical aspect of ML model development, as biased models can perpetuate structural inequalities.  The {fairmetrics} R package offers a user-friendly framework for rigorously evaluating numerous group-based fairness criteria, including metrics based on independence (e.g., statistical parity), separation (e.g., equalized odds), and sufficiency (e.g., predictive parity). Group-based fairness criteria assess whether a model is equally accurate or well-calibrated across a set of predefined groups so that appropriate bias mitigation strategies can be implemented. {fairmetrics} provides both point and interval estimates for multiple metrics through a convenient wrapper function and includes an example dataset derived from the Medical Information Mart for Intensive Care, version II (MIMIC-II) database [@goldberger2000physiobank; @raffa2016clinical].
+Fairness is a growing area of machine learning (ML) that focuses on ensuring that models do not produce systematically biased outcomes across groups defined by protected attributes, such as race, gender, or age. The {fairmetrics} R package provides a user-friendly framework for rigorously evaluating group-based fairness criteria, including independence (e.g., statistical parity), separation (e.g., equalized odds), and sufficiency (e.g., predictive parity). The package provides both point and interval estimates for a variety of commonly used criteria. {fairmetrics} also includes an example dataset derived from the Medical Information Mart for Intensive Care, version II (MIMIC-II) database [@goldberger2000physiobank; @raffa2016clinical] to demonstrate its use. 
 
 
 # Statement of Need
 
[…]
