### 2025-06-12 80c14bc
adding CRAN comments
