### 2025-06-12 a2b3fd8
re adding bibtex file
