+++ utils/joss_paper/paper.bib
@@ -1,0 +1,263 @@
+@inproceedings{lum_debias_22,
+author = {Lum, Kristian and Zhang, Yunfeng and Bower, Amanda},
+title = {De-biasing “bias” measurement},
+year = {2022},
+isbn = {9781450393522},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3531146.3533105},
+doi = {10.1145/3531146.3533105},
+abstract = {When a model’s performance differs across socially or culturally relevant groups–like race, gender, or the intersections of many such groups–it is often called ”biased.” While much of the work in algorithmic fairness over the last several years has focused on developing various definitions of model fairness (the absence of group-wise model performance disparities) and eliminating such “bias,” much less work has gone into rigorously measuring it. In practice, it important to have high quality, human digestible measures of model performance disparities and associated uncertainty quantification about them that can serve as inputs into multi-faceted decision-making processes. In this paper, we show both mathematically and through simulation that many of the metrics used to measure group-wise model performance disparities are themselves statistically biased estimators of the underlying quantities they purport to represent. We argue that this can cause misleading conclusions about the relative group-wise model performance disparities along different dimensions, especially in cases where some sensitive variables consist of categories with few members. We propose the “double-corrected” variance estimator, which provides unbiased estimates and uncertainty quantification of the variance of model performance across groups. It is conceptually simple and easily implementable without statistical software package or numerical optimization. We demonstrate the utility of this approach through simulation and show on a real dataset that while statistically biased estimators of model group-wise model performance disparities indicate statistically significant between-group model performance disparities, when accounting for statistical bias in the estimator, the estimated group-wise disparities in model performance are no longer statistically significant.},
+booktitle = {Proceedings of the 2022 ACM Conference on Fairness, Accountability, and Transparency},
+pages = {379–389},
+numpages = {11},
+location = {Seoul, Republic of Korea},
+series = {FAccT '22}
+}
+@article{obermeyerDissectingRacialBias2019,
+  title = {Dissecting Racial Bias in an Algorithm Used to Manage the Health of Populations},
+  author = {Obermeyer, Ziad and Powers, Brian and Vogeli, Christine and Mullainathan, Sendhil},
+  year = {2019},
+  journal = {Science},
+  volume = {366},
+  number = {6464},
+  pages = {447--453},
+  publisher = {American Association for the Advancement of Science}
+}
+
+@misc{mattuMachineBias,
+  title = {Machine {{Bias}}},
+  author = {Mattu, Jeff Larson,Lauren Kirchner,Surya, Julia Angwin},
+  journal = {ProPublica},
+  year = {2016},
+  urldate = {2025-05-29},
+  abstract = {There's software used across the country to predict future criminals. And it's biased against blacks.},
+  howpublished = {https://www.propublica.org/article/machine-bias-risk-assessments-in-criminal-sentencing},
+  langid = {english}
+}
+
+@article{mehrabi_survey_21,
+author = {Mehrabi, Ninareh and Morstatter, Fred and Saxena, Nripsuta and Lerman, Kristina and Galstyan, Aram},
+title = {A Survey on Bias and Fairness in Machine Learning},
+year = {2021},
+issue_date = {July 2022},
[…]
