### 2025-06-12 e24a2a6
Deploying to gh-pages from @ jianhuig/fairmetrics@92ee9e765d8f93be06ed13ea9e24ed37ec54f067 🚀
