### 2025-06-24 #8351 @Nikoleta-v3
Not sure why, but the fact that you closed one of my handling submissions and got me a new one within an hour made me laugh 😆  Sure! I'd be happy to 😄
