### 2025-06-24 #8351 @Nikoleta-v3
Hey @benyamindsmith 👋🏻 Thank you very much for your submission!

As you can see (https://github.com/openjournals/joss-reviews/issues/8351#issuecomment-2998395285), some DOIs are missing from your references—could you please update those? I also see that you've already found our list of reviewers. Could you suggest four possible reviewers by pasting their usernames in a comment? No need to ping them!

In the meantime, I'll have an initial look at the submission and will raise any issues I come across 😄
