### 2025-06-24 cd16241
addng missing dois, nothing for mlr3fairness (that I know of)
