+++ utils/joss_paper/paper.bib
@@ -22,7 +22,8 @@ series = {FAccT '22}
   volume = {366},
   number = {6464},
   pages = {447--453},
-  publisher = {American Association for the Advancement of Science}
+  publisher = {American Association for the Advancement of Science},
+  url = {https://doi.org/10.1126/science.aax2342}
 }
 
 @misc{mattuMachineBias,
@@ -223,8 +224,8 @@ series = {ITCS '12}
 
 @misc{mit_license,
       author = {Open Source Initiative},
-      title = {The MIT license}, 
-      url={https://opensource.org/license/mit}, 
+      title = {The MIT license},
+      url={https://opensource.org/license/mit},
       journal={Open Source Initiative} }
 
 
@@ -238,19 +239,19 @@ series = {ITCS '12}
   }
 
   @misc{
-  r4ds_book, 
-  title={10 Tibbles | R for Data Science}, 
-  url={https://r4ds.had.co.nz/tibbles.html}, 
-  author={Grolemund, Hadley Wickham and Garrett} 
+  r4ds_book,
+  title={10 Tibbles | R for Data Science},
+  url={https://r4ds.had.co.nz/tibbles.html},
+  author={Grolemund, Hadley Wickham and Garrett}
   }
 
   @misc{
-  fairlearn_paper, 
-  title={FairLearn: Assessing and Improving Fairness of AI Systems}, 
-  url={https://arxiv.org/abs/2303.16626}, 
-  journal={arXiv.org}, 
-  author={Weerts, Hilde and Dudík, Miroslav and Edgar, Richard and Jalali, Adrin and Lutz, Roman and Madaio, Michael}, 
-  year={2023}, month=mar 
+  fairlearn_paper,
+  title={FairLearn: Assessing and Improving Fairness of AI Systems},
+  url={https://arxiv.org/abs/2303.16626},
+  journal={arXiv.org},
+  author={Weerts, Hilde and Dudík, Miroslav and Edgar, Richard and Jalali, Adrin and Lutz, Roman and Madaio, Michael},
+  year={2023}, month=mar
   }
 
   @Manual{
