+++ utils/joss_paper/paper.bib
@@ -235,7 +235,7 @@ series = {ITCS '12}
     author = {Hadley Wickham and Jim Hester and Winston Chang and Jennifer Bryan},
     year = {2022},
     note = {R package version 2.4.5},
-    url = {https://CRAN.R-project.org/package=devtools},
+    url = {https://doi.org/10.32614/CRAN.package.devtools},
   }
 
   @misc{
@@ -260,5 +260,5 @@ series = {ITCS '12}
   author = {Florian Pfisterer and Wei Siyi and Michel Lang},
   year = {2024},
   note = {R package version 0.3.2, https://github.com/mlr-org/mlr3fairness},
-  url = {https://mlr3fairness.mlr-org.com},
+  url = {https://doi.org/10.32614/CRAN.package.mlr3fairness},
 }
