+++ utils/joss_paper/paper.bib
@@ -23,7 +23,7 @@ series = {FAccT '22}
   number = {6464},
   pages = {447--453},
   publisher = {American Association for the Advancement of Science},
-  url = {https://doi.org/10.1126/science.aax2342}
+  doi = {10.1126/science.aax2342}
 }
 
 @misc{mattuMachineBias,
@@ -235,7 +235,8 @@ series = {ITCS '12}
     author = {Hadley Wickham and Jim Hester and Winston Chang and Jennifer Bryan},
     year = {2022},
     note = {R package version 2.4.5},
-    url = {https://doi.org/10.32614/CRAN.package.devtools},
+    url = {https://CRAN.R-project.org/package=devtools},
+    doi = {10.32614/cran.package.devtools}
   }
 
   @misc{
@@ -260,5 +261,6 @@ series = {ITCS '12}
   author = {Florian Pfisterer and Wei Siyi and Michel Lang},
   year = {2024},
   note = {R package version 0.3.2, https://github.com/mlr-org/mlr3fairness},
-  url = {https://doi.org/10.32614/CRAN.package.mlr3fairness},
+  url = {https://mlr3fairness.mlr-org.com},
+  doi = {10.32614/cran.package.mlr3fairness}
 }
