+++ utils/joss_paper/paper.bib
@@ -122,7 +122,7 @@ keywords = {representation learning, natural language processing, machine learni
 @article{Fazelpour_Danks_2021, title={Algorithmic bias: Senses, sources, solutions}, volume={16}, url={https://compass.onlinelibrary.wiley.com/doi/full/10.1111/phc3.12760}, DOI={10.1111/phc3.12760}, number={8}, journal={Philosophy Compass}, author={Fazelpour, Sina and Danks, David}, year={2021}, month=jun }
 
 
-@misc{Gao_Chou_McCaw_Thurston_Varghese_Hong_Gronsbell_2024, title={What is Fair? Defining Fairness in Machine Learning for Health}, url={https://arxiv.org/abs/2406.09307}, journal={arXiv.org}, author={Gao, Jianhui and Chou, Benson and McCaw, Zachary R. and Thurston, Hilary and Varghese, Paul and Hong, Chuan and Gronsbell, Jessica}, year={2024}, month=jun }
+@misc{Gao_Chou_McCaw_Thurston_Varghese_Hong_Gronsbell_2024, title={What is Fair? Defining Fairness in Machine Learning for Health}, url={https://arxiv.org/abs/2406.09307}, journal={arXiv.org}, author={Gao, Jianhui and Chou, Benson and McCaw, Zachary R. and Thurston, Hilary and Varghese, Paul and Hong, Chuan and Gronsbell, Jessica}, year={2024}, month=jun, doi = {10.48550/arXiv.2406.09307} }
 
 
 @misc{Yfantidou_Constantinides_Spathis_Vakali_Quercia_Kawsar_2023, title={Beyond Accuracy: A Critical Review of Fairness in Machine Learning for mobile and Wearable computing}, url={https://arxiv.org/abs/2303.15585}, journal={arXiv.org}, author={Yfantidou, Sofia and Constantinides, Marios and Spathis, Dimitris and Vakali, Athena and Quercia, Daniele and Kawsar, Fahim}, year={2023}, month=mar }
@@ -150,7 +150,7 @@ series = {ITCS '12}
 
 @article{Anderson_Visweswaran_2024, title={Algorithmic individual fairness and healthcare: a scoping review}, volume={8}, url={https://academic.oup.com/jamiaopen/article/8/1/ooae149/7934945}, DOI={10.1093/jamiaopen/ooae149}, number={1}, journal={JAMIA Open}, author={Anderson, Joshua W and Visweswaran, Shyam}, year={2024}, month=dec }
 
-@misc{Plecko_Bareinboim_2022, title={Causal Fairness analysis}, url={https://arxiv.org/abs/2207.11385}, journal={arXiv.org}, author={Plecko, Drago and Bareinboim, Elias}, year={2022}, month=jul }
+@misc{Plecko_Bareinboim_2022, title={Causal Fairness analysis}, url={https://arxiv.org/abs/2207.11385}, journal={arXiv.org}, author={Plecko, Drago and Bareinboim, Elias}, year={2022}, month=jul, doi = {10.48550/arXiv.2207.11385} }
 
 @article{Makhlouf_Zhioua_Palamidessi_2024, title={When causality meets fairness: A survey}, volume={141}, url={https://www.sciencedirect.com/science/article/abs/pii/S2352220824000543}, DOI={10.1016/j.jlamp.2024.101000}, journal={Journal of Logical and Algebraic Methods in Programming}, author={Makhlouf, Karima and Zhioua, Sami and Palamidessi, Catuscia}, year={2024}, month=jun, pages={101000} }
 
@@ -177,6 +177,7 @@ series = {ITCS '12}
     year = {2021},
[…]
