### 2025-06-24 8e821c2
adding all dois
