+++ utils/joss_paper/paper.bib
@@ -125,7 +125,7 @@ keywords = {representation learning, natural language processing, machine learni
 @misc{Gao_Chou_McCaw_Thurston_Varghese_Hong_Gronsbell_2024, title={What is Fair? Defining Fairness in Machine Learning for Health}, url={https://arxiv.org/abs/2406.09307}, journal={arXiv.org}, author={Gao, Jianhui and Chou, Benson and McCaw, Zachary R. and Thurston, Hilary and Varghese, Paul and Hong, Chuan and Gronsbell, Jessica}, year={2024}, month=jun, doi = {10.48550/arXiv.2406.09307} }
 
 
-@misc{Yfantidou_Constantinides_Spathis_Vakali_Quercia_Kawsar_2023, title={Beyond Accuracy: A Critical Review of Fairness in Machine Learning for mobile and Wearable computing}, url={https://arxiv.org/abs/2303.15585}, journal={arXiv.org}, author={Yfantidou, Sofia and Constantinides, Marios and Spathis, Dimitris and Vakali, Athena and Quercia, Daniele and Kawsar, Fahim}, year={2023}, month=mar }
+@misc{Yfantidou_Constantinides_Spathis_Vakali_Quercia_Kawsar_2023, title={Beyond Accuracy: A Critical Review of Fairness in Machine Learning for mobile and Wearable computing}, url={https://arxiv.org/abs/2303.15585}, journal={arXiv.org}, author={Yfantidou, Sofia and Constantinides, Marios and Spathis, Dimitris and Vakali, Athena and Quercia, Daniele and Kawsar, Fahim}, year={2023}, month=mar, doi = {10.48550/arXiv.2303.15585} }
 
 @article{Castelnovo_Crupi_Greco_Regoli_Penco_Cosentini_2022, title={A clarification of the nuances in the fairness metrics landscape}, volume={12}, url={https://www.nature.com/articles/s41598-022-07939-1}, DOI={10.1038/s41598-022-07939-1}, number={1}, journal={Scientific Reports}, author={Castelnovo, Alessandro and Crupi, Riccardo and Greco, Greta and Regoli, Daniele and Penco, Ilaria Giuseppina and Cosentini, Andrea Claudio}, year={2022}, month=mar }
 
@@ -154,7 +154,7 @@ series = {ITCS '12}
 
 @article{Makhlouf_Zhioua_Palamidessi_2024, title={When causality meets fairness: A survey}, volume={141}, url={https://www.sciencedirect.com/science/article/abs/pii/S2352220824000543}, DOI={10.1016/j.jlamp.2024.101000}, journal={Journal of Logical and Algebraic Methods in Programming}, author={Makhlouf, Karima and Zhioua, Sami and Palamidessi, Catuscia}, year={2024}, month=jun, pages={101000} }
 
-@misc{Awasthi_Cortes_Mansour_Mohri_2020, title={Beyond individual and group fairness}, url={https://arxiv.org/abs/2008.09490}, journal={arXiv.org}, author={Awasthi, Pranjal and Cortes, Corinna and Mansour, Yishay and Mohri, Mehryar}, year={2020}, month=aug }
+@misc{Awasthi_Cortes_Mansour_Mohri_2020, title={Beyond individual and group fairness}, url={https://arxiv.org/abs/2008.09490}, journal={arXiv.org}, author={Awasthi, Pranjal and Cortes, Corinna and Mansour, Yishay and Mohri, Mehryar}, year={2020}, month=aug, doi = {10.48550/arXiv.2008.09490} }
 
[…]
