### 2025-06-24 #8351 @benyamindsmith
@Nikoleta-v3  thank you for agreeing to serve as our editor.

1. I just handled the missing DOIs. in the references and hunted down other references which have DOIs available.

2. Looking at the reviewers I see the following editors who I think may be relevant:

- [VasanthRajendran](https://github.com/VasanthRajendran) - he actually reached out asking to review the paper earlier.
- [vankesteren](https://github.com/vankesteren)
- [StevenGolovkine](https://github.com/StevenGolovkine)
- [mohitcek](https://github.com/mohitcek)
- [fernandosola](https://github.com/fernandosola)

or anyone with experience with model diagnostics I believe would be suitable.
