### 2025-06-24 #8351 @Nikoleta-v3
@vankesteren & @StevenGolovkine 👋🏻 Would you be willing to review this submission for JOSS?

As you already know we conduct our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html 😄

The submission I'd like you to review is titled: "fairmetrics: An R package for group fairness evaluation". You can find more details at the top of this GitHub issue 🆙

Please let me know if you're available 😄 Since the Easter break is coming up in some countries, we can delay the review deadline. Thank you!
