### 2025-06-26 #8351 @StevenGolovkine
Thanks for the invitation. I'll be happy to review it.
