### 2025-06-27 #8497 @editorialbot
[REVIEW]: fairmetrics: An R package for group fairness evaluation
**Submitting author:** @benyamindsmith (<a href="http://orcid.org/0009-0007-2206-0177">Benjamin Smith</a>)
**Repository:** https://github.com/jianhuig/fairmetrics
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.2
