### 2025-07-02 jianhuig/fairmetrics#15 @StevenGolovkine
Issue with the General Checks
Hi @benyamindsmith. I want first to congratulate you and the other persons that have been working on the package and the submission to JOSS.

I'll create several issues (one for each part of the JOSS review checklist) to group my comments with respect to the submitted paper, and in relation with the open review in [openjournals/joss-reviews/issues/8497](https://github.com/openjournals/joss-reviews/issues/8497).

1. There are two LICENSE files in your repo. Could you please delete one of them.

2. It seems that @Benson-chou did some work on the package. However, he/she is not listed as author of the manuscript, could you expand on this ?
