### 2025-07-02 jianhuig/fairmetrics#16 @StevenGolovkine
Issue with Functionality
Comments related to the Functionality part of the JOSS review checklist ([openjournals/joss-reviews/issues/8497](https://github.com/openjournals/joss-reviews/issues/8497)).

1. You claim a huge gain in memory usage when using your package. How did you compute these numbers?

2. Have you done some comparison regarding the computational time between your package and others?
