### 2025-07-02 jianhuig/fairmetrics#15 @benyamindsmith
Hi @StevenGolovkine 👋

1. I tried to remove one of them but when I run `devtools::check()` I get a warning thrown about its submit ability to CRAN. I am using `usethis::use_mit_license()` for generating the license.

2. For this question I'll defer to my advisor @jlgrons  and co-author @jianhuig. Initially he was listed as an author, however, I believe the reason for this is because Benson was not involved in the writing of the article, which was why he was not included an an author for the article. He is however listed as an author on the CRAN package.

Let me know if you have any other questions.
