### 2025-07-05 jianhuig/fairmetrics#16 @StevenGolovkine
1. Thanks for the code. I ran it on my laptop and find similar numbers.
2. I agree that computational time is not especially relevant for the comparison with other software. However, from a general perspective, I think it is important as if your code takes forever to run, no one will use it. Considering your code, the bootstrapping part takes most of the computational time (if I get it well). As a suggestion for a future version of the code, you can try to optimize this part, for example, using multi-threading and parallelism.

You can close this issue now.
