### 2025-07-06 jianhuig/fairmetrics#16 @benyamindsmith
Thank you @StevenGolovkine

Regarding your second point, accommodating multi-threading/parallelism an important point and definitely something worth addressing in a later version of this package.

I have opened #17  to note this feature request  for a future version.
