### 2025-07-07 jianhuig/fairmetrics#18 @StevenGolovkine
Issue with the Documentation
Comments related to the Documentation part of the JOSS review checklist [openjournals/joss-reviews/issues/8497](https://github.com/openjournals/joss-reviews/issues/8497).

1. It would be great to add a statement of need in the documentation in order to help users of the package to understand why the software exists without seeing the paper. You can copy/paste the statement of need from the paper.
2. There is a Latex math issue in the description of the functions `eval_neg_class_bal` and `eval_pos_class_bal`.
3. Concerning the tests, it seems that you only test if the code runs without errors. Could you also test whether you get expected results for a given input?
4. Can you add a CONTRIBUTING file in the repo and a page in the documentation where you explain the guidelines for someone that would like to contribute to the software? It should also describe how to report issues/problems with the software and how to seek support.
