### 2025-07-08 jianhuig/fairmetrics#19 @StevenGolovkine
Issue with the Software paper
Comments related to the Software paper part of the JOSS review checklist [openjournals/joss-reviews/issues/8497](https://github.com/openjournals/joss-reviews/issues/8497).

The paper is well written, I only have few comments.
1. Remove the curly brackets around the packages name. Please use fairmetrics (without particular font) in the title and `fairmetrics` in the text to conform to others JOSS papers.
2. The citation Mattu(2016) (l.16) should be in parenthesis.
3. Consider to use `data.frame` and `tibble` in line 40 as it represents variable type and not noun.
4. The references have to be cleaned. Lots of the them are included in the References section but not in the main text.
