### 2025-07-18 6450872
fixing latex issues with eval_neg_class_bal and eval_pos_class_bal
