+++ man/eval_pos_class_bal.Rd
@@ -53,7 +53,7 @@ confidence interval for the ratio in average predicted probability
 \description{
 This function evaluates \emph{Balance for the Positive Class}, a fairness criterion
 that checks whether the model assigns similar predicted probabilities across groups
-among individuals whose true outcome is positive (i.e., \(Y = 1\)).
+among individuals whose true outcome is positive (i.e. \eqn{Y = 1}).
 }
 \examples{
 \donttest{
