+++ cran-comments.md
@@ -2,6 +2,11 @@
 
 The code here has been reformatted and edited for release with a associated article in the Journal of Open Source Software. I am in contact with the Journal Editors and they will be reviewing this reviewing this version of the package.
 
+
+## Version 1.0.4
+
+* Implementing Edits Required by JOSS TEAM
+
 ## Version 1.0.3
 
 * Changing abbreviations to full names for `eval_*` functions. 
