### 2025-07-18 bd08139
adding new tests to ensure eval_* functions work as expected
