### 2025-07-18 9c8a73e
Added Contributing.md file
