### 2025-07-18 04c16ed
editing paper to match reviewer suggestions. Holding off on references for now until clarification
