### 2025-07-18 jianhuig/fairmetrics#19 @benyamindsmith
Hello @StevenGolovkine

1 - 3. Done (04c16ed)
4.  The references listed here consist of many from the manuscript which the paper is based on. In particular, we are interested in ensuring that [this paper](https://arxiv.org/abs/2406.09307) is cited.  The reason for all the references is based on the specification in the YAML header to do that.

https://github.com/jianhuig/fairmetrics/blob/04c16ed4279f9d1e4fe2bc533e170d5bdc427217/utils/joss_paper/paper.md?plain=1#L23-L24

Let me know how to proceed.

Tagging @jlgrons and @jianhuig to keep them in the loop about point 4.
