### 2025-07-18 jianhuig/fairmetrics#19 @StevenGolovkine
If you want to cite your paper, make sure to use it in your main text. Otherwise, no paper should be in the References section if it is not in cite in the main text. You should probably not use the `nocite` YAML header.
