### 2025-07-18 jianhuig/fairmetrics#18 @StevenGolovkine
I've been thinking about the statement of need. Actually, I change my mind. No need to add it to the documentation, because it is already in the description.

Concerning the test, I think I was not clear. I am sure that the `get_fairness_metrics` works in your test, because it is like running the same code twice (especially, as you don't ask for confidence intervals in your tests).
I was thinking about testing the output for a given numerical input. For example, for pathological cases, e.g. the model predict 1 for every observation, you should be able to compute the true value of each metric. Do your functions return this value with the right confidence interval? Because of the randomness of the bootstrap you may want to bound the difference between the true value and the one that is return by the functions. You may also think of other cases that are derivable.
