### 2025-07-21 b5b9b7a
cleaning up references, citing Gao 2024)
