### 2025-07-21 jianhuig/fairmetrics#19 @benyamindsmith
@StevenGolovkine good to know.

I got rid of `nocite` in the YAML header and cited the paper in the statement of need (b5b9b7a).

Let me know if there's anything else that needs to be addressed in terms of the paper's content.
