### 2025-07-21 d745036
removing sentence, keeping citation
