### 2025-07-21 jianhuig/fairmetrics#18 @benyamindsmith
@StevenGolovkine got it.

I came up with 71f0cbf . Let me know if this is something that you have more in mind for the tests.
