### 2025-07-21 jianhuig/fairmetrics#18 @StevenGolovkine
By looking at your test file, sometimes you test for `NA`, sometimes for `NaN`, is it the expected behavior to get `NA` or `NaN`? Also, I try this:
`result <- eval_eq_opp(data = dt, outcome = "y_true", group = "group", probs = "y_pred", digits = 1, confint = TRUE, message = TRUE)` and got this error message: `Error in if (lower_ci > 0 | upper_ci < 0) { :
  missing value where TRUE/FALSE needed`, is this the intended behavior as all the arguments seem legit? (I did not try every function, but I guess I'll get the same results)
