### 2025-07-21 jianhuig/fairmetrics#18 @benyamindsmith
@StevenGolovkine good catch. I will look at fixing this behavior and get back to you on this.
