### 2025-07-22 2c0d3fc
adding test, fixing message structure, documentation
