+++ R/Fairness.R
@@ -25,7 +25,7 @@
 #' @param digits Integer indicating the number of decimal places to round
 #' results to; defaults to 2.
 #' @param message Logical; if TRUE (default), prints a textual summary of the
-#' fairness evaluation.
+#' fairness evaluation. Only works if `confint` is TRUE.
 #' @return A data frame summarizing FNR-based group disparity metrics with the
 #' following columns:
 #'
@@ -137,6 +137,16 @@ eval_eq_opp <- function(data, outcome, group, probs, cutoff = 0.5, confint = TRU
       paste0("Group", sort(unique(data[[group]]))[[2]]),
       "Difference", "95% Diff CI", "Ratio", "95% Ratio CI"
     )
+
+    # Print message if desired
+    if (message) {
+      if (lower_ci > 0 | upper_ci < 0) {
+        cat("There is evidence that model does not satisfy equal opportunity.\n")
+      } else {
+        cat("There is not enough evidence that the model does not satisfy equal
+          opportunity.\n")
+      }
+    }
   } else{
     results_df <- data.frame(
       "False Negative Rate",
@@ -153,15 +163,7 @@ eval_eq_opp <- function(data, outcome, group, probs, cutoff = 0.5, confint = TRU
     )
   }
 
-  # Print message if desired
-  if (message) {
-    if (lower_ci > 0 | upper_ci < 0) {
-      cat("There is evidence that model does not satisfy equal opportunity.\n")
-    } else {
-      cat("There is not enough evidence that the model does not satisfy equal
-          opportunity.\n")
-    }
-  }
+
 
   return(results_df)
 }
@@ -194,7 +196,7 @@ eval_eq_opp <- function(data, outcome, group, probs, cutoff = 0.5, confint = TRU
 #' @param digits Number of decimal places to round numeric results; defaults to
 #' 2.
 #' @param message Logical; if TRUE (default), prints a textual summary of the
-#' fairness evaluation.
+#' fairness evaluation. Only works if `confint` is TRUE.
 #' @return A data frame summarizing group disparities in both FNR and FPR with
 #' the following columns:
 #' \itemize{
@@ -334,6 +336,17 @@ eval_eq_odds <- function(data, outcome, group, probs, cutoff = 0.5, confint = TR
       paste0("Group ", sort(unique(data[[group]]))[[2]]),
       "Difference", "95% CI", "Ratio", "95% CI"
     )
+
+    # Print summary message if desired
+    if (message) {
+      if (any(fnr_lower > 0) || any(fnr_upper < 0) || any(fpr_lower > 0) ||
+          any(fpr_upper < 0)) {
+        cat("There is evidence that model does not satisfy equalized odds.\n")
+      } else {
+        cat("There is not enough evidence that the model does not satisfy the
+          equalized odds criterion.\n")
+      }
+    }
   }else{
 
     # Structure the results as a dataframe
@@ -353,16 +366,7 @@ eval_eq_odds <- function(data, outcome, group, probs, cutoff = 0.5, confint = TR
     )
   }
 
-  # Print summary message if desired
-  if (message) {
-    if (any(fnr_lower > 0) || any(fnr_upper < 0) || any(fpr_lower > 0) ||
-      any(fpr_upper < 0)) {
-      cat("There is evidence that model does not satisfy equalized odds.\n")
-    } else {
-      cat("There is not enough evidence that the model does not satisfy the
-          equalized odds criterion.\n")
-    }
-  }
+
 
   return(results_df)
 }
@@ -382,7 +386,8 @@ eval_eq_odds <- function(data, outcome, group, probs, cutoff = 0.5, confint = TR
 #' @param bootstraps Number of bootstrap samples, default is 2500
[…]
