### 2025-07-22 88a6e52
applying na.rm=TRUE for diff sd()
