### 2025-07-22 ed147b6
fixing NA outcomes
