+++ R/Fairness.R
@@ -117,9 +117,9 @@ eval_eq_opp <- function(data, outcome, group, probs, cutoff = 0.5, confint = TRU
     lower_ci <- round(fnr_diff - qnorm(1 - alpha / 2) * sd(se[1, ], na.rm = TRUE), digits)
     upper_ci <- round(fnr_diff + qnorm(1 - alpha / 2) * sd(se[1, ], na.rm = TRUE), digits)
     lower_ratio_ci <- round(exp(log(fnr_ratio) - qnorm(1 - alpha / 2) *
-                                  sd(se[2, ])), digits)
+                                  sd(se[2, ], na.rm=TRUE)), digits)
     upper_ratio_ci <- round(exp(log(fnr_ratio) + qnorm(1 - alpha / 2) *
-                                  sd(se[2, ])), digits)
+                                  sd(se[2, ], na.rm=TRUE)), digits)
 
     # Create a dataframe for the results
     results_df <- data.frame(
@@ -298,18 +298,18 @@ eval_eq_odds <- function(data, outcome, group, probs, cutoff = 0.5, confint = TR
     })
 
     # Calculate confidence intervals
-    fnr_lower <- round(fnr_diff - qnorm(1 - alpha / 4) * sd(se[1, ]), digits)
-    fnr_upper <- round(fnr_diff + qnorm(1 - alpha / 4) * sd(se[1, ]), digits)
-    fpr_lower <- round(fpr_diff - qnorm(1 - alpha / 4) * sd(se[2, ]), digits)
-    fpr_upper <- round(fpr_diff + qnorm(1 - alpha / 4) * sd(se[2, ]), digits)
+    fnr_lower <- round(fnr_diff - qnorm(1 - alpha / 4) * sd(se[1, ],na.rm=TRUE), digits)
+    fnr_upper <- round(fnr_diff + qnorm(1 - alpha / 4) * sd(se[1, ],na.rm=TRUE), digits)
+    fpr_lower <- round(fpr_diff - qnorm(1 - alpha / 4) * sd(se[2, ],na.rm=TRUE), digits)
+    fpr_upper <- round(fpr_diff + qnorm(1 - alpha / 4) * sd(se[2, ],na.rm=TRUE), digits)
     fnr_ratio_lower <- round(exp(log(fnr_ratio) - qnorm(1 - alpha / 4) *
-                                   sd(se[3, ])), digits)
+                                   sd(se[3, ],na.rm=TRUE)), digits)
     fnr_ratio_upper <- round(exp(log(fnr_ratio) + qnorm(1 - alpha / 4) *
-                                   sd(se[3, ])), digits)
+                                   sd(se[3, ],na.rm=TRUE)), digits)
     fpr_ratio_lower <- round(exp(log(fpr_ratio) - qnorm(1 - alpha / 4) *
-                                   sd(se[4, ])), digits)
+                                   sd(se[4, ],na.rm=TRUE)), digits)
     fpr_ratio_upper <- round(exp(log(fpr_ratio) + qnorm(1 - alpha / 4) *
-                                   sd(se[4, ])), digits)
+                                   sd(se[4, ],na.rm=TRUE)), digits)
 
     # Structure the results as a dataframe
     results_df <- data.frame(
[…]
