+++ tests/testthat/test-Fairness-outcomes.R
@@ -76,6 +76,48 @@ test_that('eval_bs_parity CI outcome test', {
 })
 
 
+test_that('eval_eq_odds ratio+diff outcome test', {
+  expect_equal({
+    dt <- data.frame(
+      y_true = c(0, 1, 0, 1),
+      y_pred = c(1, 1, 1, 1),
+      group = c("A", "A", "B", "B")
+    )
+    result <- eval_eq_odds(data = dt,
+                          outcome = "y_true",
+                          group = "group",
+                          probs = "y_pred",
+                          digits = 1,
+                          confint = TRUE,
+                          message = TRUE)
+    list(result$Ratio, result$Difference)
+  },
+  list(
+    "NaN; 1",
+    "0; 0"
+    ))
+})
+
+test_that('eval_eq_odds CI outcome test', {
+  expect_equal({
+    dt <- data.frame(
+      y_true = c(0, 1, 0, 1),
+      y_pred = c(1,1,1,1),
+      group = c("A", "A", "B", "B")
+    )
+    result <- eval_eq_odds(data = dt,
+                          outcome = "y_true",
+                          group = "group",
+                          probs = "y_pred",
+                          digits = 1,
+                          confint = TRUE,
+                          message = TRUE)
+    list(result$`95% Ratio CI`, result$`95% Diff CI`)
+  },
+  list("[NaN, NaN]; [1, 1]","[0, 0]; [0, 0]"))
+})
+
+
 test_that('eval_eq_opp ratio+diff outcome test', {
   expect_equal({
     dt <- data.frame(
@@ -84,12 +126,12 @@ test_that('eval_eq_opp ratio+diff outcome test', {
       group = c("A", "A", "B", "B")
     )
     result <- eval_eq_opp(data = dt,
-                             outcome = "y_true",
-                             group = "group",
-                             probs = "y_pred",
-                             digits = 1,
-                             confint = TRUE,
-                             message = TRUE)
+                          outcome = "y_true",
+                          group = "group",
+                          probs = "y_pred",
+                          digits = 1,
+                          confint = TRUE,
+                          message = TRUE)
     list(result$Ratio, result$Difference)
   },
   list(NaN,0))
@@ -103,17 +145,16 @@ test_that('eval_eq_opp CI outcome test', {
       group = c("A", "A", "B", "B")
     )
     result <- eval_eq_opp(data = dt,
-                             outcome = "y_true",
-                             group = "group",
-                             probs = "y_pred",
-                             digits = 1,
-                             confint = TRUE,
-                             message = TRUE)
+                          outcome = "y_true",
+                          group = "group",
+                          probs = "y_pred",
+                          digits = 1,
+                          confint = TRUE,
+                          message = TRUE)
     list(result$`95% Ratio CI`, result$`95% Diff CI`)
   },
   list("[NaN, NaN]", "[0, 0]"))
 })
-
 
 
 test_that('eval_pred_equality ratio+diff outcome test', {
@@ -124,16 +165,17 @@ test_that('eval_pred_equality ratio+diff outcome test', {
       group = c("A", "A", "B", "B")
     )
     result <- eval_pred_equality(data = dt,
-                          outcome = "y_true",
-                          group = "group",
-                          probs = "y_pred",
-                          digits = 1,
-                          confint = TRUE,
-                          message = TRUE)
+                                 outcome = "y_true",
+                                 group = "group",
+                                 probs = "y_pred",
+                                 digits = 1,
+                                 confint = TRUE,
+                                 message = TRUE)
     list(result$Ratio, result$Difference)
   },
   list(1,0))
 })
+
 
 test_that('eval_pred_equality CI outcome test', {
   expect_equal({
