### 2025-07-22 jianhuig/fairmetrics#18 @benyamindsmith
@StevenGolovkine I believe I cleaned things up accordingly.

1. The ratio returning NA/NaN is expected at `0/0` is undefined. Based on the tests which I've updated, `NaN` is now used.
2. The issue with the `message` argument is that it was conditional on the `confint` argument (which it needs to be in order to print a statement).  I addressed this issue across all functions . i.e. `result <- eval_eq_opp(data = dt, outcome = "y_true", group = "group", probs = "y_pred", digits = 1, confint = TRUE, message = TRUE)` (and other functions) should not return an error anymore.

Let me know if you have any other feedbacks.  If not, would you like me to publish these changes to CRAN before publication of the manuscript?
