### 2025-07-22 #8497 @StevenGolovkine
Hi @Nikoleta-v3,

The review is finished for me, the authors answered to all my comments.
Just one point regarding [https://github.com/jianhuig/fairmetrics/issues/15](https://github.com/jianhuig/fairmetrics/issues/15), one of the contributors to the package is not listed as an author of the paper. I let you decide if the authors should be changed.
