### 2025-07-25 #8497 @Nikoleta-v3
Thank you very much for all your work, @StevenGolovkine! I’ve seen your concern, and my apologies to both you and @benyamindsmith for the delay in replying. After consulting with the other editors, I’d like to summarize where we stand:

From what I understand, @Benson-chou contributed significantly to the development of the package. Given this, they would qualify for authorship. The fact that they did not contribute to the writing of the paper may simply be because they were never asked.

To clarify the situation, @benyamindsmith should reach out to @Benson-chou—whose authorship status is in question—and provide a written response in the review thread confirming that the person has either been added as a coauthor or has formally declined to be listed.
