### 2025-07-27 #8497 @benyamindsmith
Hello @Nikoleta-v3

We have added Benson as a co-author to the paper after speaking with him.

Our apologies for the misunderstanding, we were under impression there had to be a contribution to the article to be considered an author.

That being said, Benson did contribute to an earlier version of the code.
