### 2025-07-27 jianhuig/fairmetrics#15 @benyamindsmith
__Update:__ Correspondence continuing here: openjournals/joss-reviews/issues/8497

@StevenGolovkine could I close this issue?
