### 2025-07-28 jianhuig/fairmetrics#23 @vankesteren
JOSS review [Documentation]: Clean up vignette tutorial
The vignette [here](https://jianhuig.github.io/fairmetrics/articles/fairmetrics.html) can be much cleaner, which would be better for new users. As a main goal, I'd expect this tutorial to explain __how to obtain, interpret, and report the fairness metrics__ using as simple an example as possible.

That means, for example:
- Don't go into data cleaning, start with the preprocessed version
- Don't focus too much on model building
- train-test split is important to include
- Load your package, don't call it using `fairmetrics::function()`
- Hide the distracting `kableExtra`code
