### 2025-07-28 jianhuig/fairmetrics#24 @vankesteren
JOSS review [general]: Binary sensitive attribute constraint
Your package only works for binary groups, such as binary gender, a binary race variable, or some splitted continuous variable.

This is understandable, as the metrics in this package focus on comparing various confusion matrix attributes between groups.

However, this also severely limits the usability. For example, what if I have "migration background" with different continents/regions? What if I have male/female/other as gender? These are just 2 common examples.

There is nothing technical (outside of implementation) standing in the way of extending the methods in this package to more than 2 categories (I think!), as you can compare to a reference category or do one-vs-all in your metrics.

I have 2 suggestions:
- Make it more explicit in the readme / description / paper that this package is designed for binary sensitive attributes only
- Explore / plan how to extend the package to more groups
