### 2025-07-28 jianhuig/fairmetrics#25 @vankesteren
JOSS review [general]: pitfall of statistical inference
One of the main advantages of your method is the confidence intervals / statistical inference. I believe that your approach is very sensitive to a common, well-documented pitfall of statistical inference, particularly with null hypothesis significance testing: __absence of evidence (p < 0.05) is not evidence of absence__.

I am of the opinion that __you should warn users about this very explicitly__ in your package documentation, in the tutorial, and in the paper (basically everywhere).

The language used in your vignette already hints at this pitfall (emphasis mine):

> On the other hand, the sufficiency criterion __is satisfied__ as predictive parity shows no significant difference between males and females (difference: -4%, 95% CI: [-21%, 13%]; ratio: 0.94, 95% CI: [0.72, 1.23]). This suggests that given the same prediction, the actual mortality rates are similar for both males and females.

_(NB: I am not accusing you of making this mistake, just that the commonly used language around statistical inference is prone to being interpreted wrongly)_

Instead, it should be the size of the effect in combination with the uncertainty around the effect that yields evidence for statistical parity. In your example, I think 20% would be quite a large predictive parity difference, and it is within the confidence set. With `p<0.05` we can only state that we do not have enough evidence to reject the null.

Failing to warn users about this would be irresponsible: a user with few observations in the test set will fail to reject the null for all metrics, with high probability. This does not mean that their model is fair.
