### 2025-07-28 jianhuig/fairmetrics#27 @vankesteren
JOSS review [paper / code]: duplicate row names in output / more user-friendly output
In your paper you show the following table:

Note that positive prediction rate and avg. predicted prob. appear twice. This is unclear to the user and should be fixed.

In general, I think the output can be tidier / more user-friendly. In the second table, for example, it is unclear which group (level) has a higher or lower value on each metric:

maybe something that links the two explicitly can solve this issue?

| metric | value | males | females | difference | diff ci | ratio | ratio ci |
| :----- | :---- | :---- | :------ | :--------- | :------ | :---- | :------- |
| Statistical parity | Positive prediction rate | 0.18 | 0.09 | 0.08|  [0.04, 0.12] | 2.00 | [1.41, 2.83] |
|...|...|...|...|...|...|...|...|...|
