### 2025-07-28 #8497 @vankesteren
Sorry for the delay; I have now finished a quick first round of this review and added issues in the repo. I will continue the review after these have been addressed (discussed or implemented)
