+++ R/eval_pred_equality.R
@@ -1,0 +1,152 @@
+#' Examine Predictive Equality of a Model
+#'
+#' This function evaluates predictive equality, a fairness metric that compares the
+#' False Positive Rate (FPR) between groups defined by a sensitive attribute. It assesses
+#' whether individuals from different groups are equally likely to be incorrectly flagged as
+#' positive when they are, in fact, negative.
+#'
+#' @param data Data frame containing the outcome, predicted outcome, and
+#' sensitive attribute
+#' @param outcome Name of the outcome variable, it must be binary
+#' @param group Name of the sensitive attribute
+#' @param probs Name of the predicted outcome variable
+#' @param cutoff Threshold for the predicted outcome, default is 0.5
+#' @param confint Whether to compute 95% confidence interval, default is TRUE
+#' @param bootstraps Number of bootstrap samples, default is 2500
+#' @param alpha The 1 - significance level for the confidence interval, default is 0.05
+#' @param digits Number of digits to round the results to, default is 2
+#' @param message Logical; if TRUE (default), prints a textual summary of the
+#' fairness evaluation. Only works if `confint` is TRUE.
+#' @return A list containing the following elements:
+#' - FPR_Group1: False Positive Rate for the first group
+#' - FPR_Group2: False Positive Rate for the second group
+#' - FPR_Diff: Difference in False Positive Rate
+#' - FPR_Ratio: Ratio in False Positive Rate
+#' If confidence intervals are computed (`confint = TRUE`):
+#' - FPR_Diff_CI: A vector of length 2 containing the lower and upper bounds
+#' of the 95% confidence interval for the difference in False Positive Rate
+#' - FPR_Ratio_CI: A vector of length 2 containing the lower and upper bounds
+#' of the 95% confidence interval for the ratio in False Positive Rate
+#' @importFrom stats qnorm sd
+#' @examples
+#' \donttest{
+#' library(fairmetrics)
+#' library(dplyr)
+#' library(magrittr)
+#' library(randomForest)
+#' data("mimic_preprocessed")
+#' set.seed(123)
+#' train_data <- mimic_preprocessed %>%
+#'   dplyr::filter(dplyr::row_number() <= 700)
+#' # Fit a random forest model
+#' rf_model <- randomForest::randomForest(factor(day_28_flg) ~ ., data = train_data, ntree = 1000)
+#' # Test the model on the remaining data
+#' test_data <- mimic_preprocessed %>%
+#'   dplyr::mutate(gender = ifelse(gender_num == 1, "Male", "Female")) %>%
+#'   dplyr::filter(dplyr::row_number() > 700)
+#'
+#' test_data$pred <- predict(rf_model, newdata = test_data, type = "prob")[, 2]
+#'
+#' # Fairness evaluation
+#' # We will use sex as the sensitive attribute and day_28_flg as the outcome.
+#' # We choose threshold = 0.41 so that the overall FPR is around 5%.
+#'
+#' # Evaluate Predictive Equality
+#' eval_pred_equality(
+#'   data = test_data,
+#'   outcome = "day_28_flg",
+#'   group = "gender",
+#'   probs = "pred",
+#'   cutoff = 0.41
+#' )
+#' }
+#' @seealso \code{\link{eval_pos_pred_parity}},  \code{\link{eval_neg_pred_parity}}, \code{\link{eval_stats_parity}}
+#' @export
+
[…]
