+++ R/eval_acc_parity.R
@@ -1,0 +1,152 @@
+#' Examine Accuracy Parity of a Model
+#'
+#' This function assesses *Accuracy Parity*, a fairness criterion that evaluates whether
+#' the overall accuracy of a predictive model is consistent across different groups.
+#'
+#' @param data Data frame containing the outcome, predicted outcome, and
+#' sensitive attribute
+#' @param outcome Name of the outcome variable
+#' @param group Name of the sensitive attribute
+#' @param probs Predicted probabilities
+#' @param cutoff Cutoff value for the predicted probabilities
+#' @param confint Logical indicating whether to calculate confidence intervals
+#' @param bootstraps Number of bootstraps to use for confidence intervals
+#' @param alpha The 1 - significance level for the confidence interval, default is 0.05
+#' @param digits Number of digits to round the results to, default is 2
+#' @param message Logical; if TRUE (default), prints a textual summary of the
+#' fairness evaluation. Only works if `confint` is TRUE.
+#' @return A list containing the following elements:
+#' - Accuracy for Group 1
+#' - Accuracy for Group 2
+#' - Difference in accuracy
+#' - Ratio in accuracy
+#' If confidence intervals are computed (`confint = TRUE`):
+#' - A vector of length 2 containing the lower and upper bounds of the 95%
+#' confidence interval for the difference in accuracy
+#' - A vector of length 2 containing the lower and upper bounds of the 95%
+#' confidence interval for the ratio in accurac
+#' @importFrom stats qnorm sd
+#' @examples
+#' \donttest{
+#' library(fairmetrics)
+#' library(dplyr)
+#' library(magrittr)
+#' library(randomForest)
+#' data("mimic_preprocessed")
+#' set.seed(123)
+#' train_data <- mimic_preprocessed %>%
+#'   dplyr::filter(dplyr::row_number() <= 700)
+#' # Fit a random forest model
+#' rf_model <- randomForest::randomForest(factor(day_28_flg) ~ ., data = train_data, ntree = 1000)
+#' # Test the model on the remaining data
+#' test_data <- mimic_preprocessed %>%
+#'   dplyr::mutate(gender = ifelse(gender_num == 1, "Male", "Female")) %>%
+#'   dplyr::filter(dplyr::row_number() > 700)
+#'
+#' test_data$pred <- predict(rf_model, newdata = test_data, type = "prob")[, 2]
+#'
+#' # Fairness evaluation
+#' # We will use sex as the sensitive attribute and day_28_flg as the outcome.
+#' # We choose threshold = 0.41 so that the overall FPR is around 5%.
+#'
+#' # Evaluate Accuracy Parity
+#' eval_acc_parity(
+#'   data = test_data,
+#'   outcome = "day_28_flg",
+#'   group = "gender",
+#'   probs = "pred",
+#'   cutoff = 0.41
+#' )
+#' }
+#'
+#' @seealso \code{\link{eval_cond_acc_equality}}
+#' @export
+
+eval_acc_parity <- function(data, outcome, group, probs, cutoff = 0.5, confint = TRUE,
+                            alpha = 0.05, bootstraps = 2500,
+                            digits = 2, message = TRUE) {
+  # Check if outcome is binary
+  unique_values <- unique(data[[outcome]])
+  if (!(length(unique_values) == 2 && all(unique_values %in% c(0, 1)))) {
+    stop("Outcome must be binary (containing only 0 and 1).")
+  }
+
+  acc <- get_acc(
[…]
