### 2025-07-29 #8497 @Nikoleta-v3
Replying to: https://github.com/openjournals/joss-reviews/issues/8497#issuecomment-3123836018.

Thank you very much, @benyamindsmith, for clearing up the situation! Happy to see that Benson-chou is now an author.

And thank you to @StevenGolovkine for bringing it to our attention 😄 🙏🏻
