### 2025-07-29 e521334
going to edit get_fairness_metrics, this is a security commit that I can revert to
