### 2025-07-30 cf5c175
refactored get_fairness metrics to address bootstrap inefficiency
