+++ R/get_fairness_metrics.R
@@ -83,200 +83,593 @@
 #'
 #' @export
 get_fairness_metrics <- function(data,
-                                 outcome,
-                                 group,
-                                 group2 = NULL,
-                                 condition = NULL,
-                                 probs,
-                                 confint = TRUE,
-                                 cutoff = 0.5,
-                                 bootstraps = 2500,
-                                 alpha = 0.05,
-                                 digits = 2) {
+                                  outcome,
+                                  group,
+                                  group2 = NULL,
+                                  condition = NULL,
+                                  probs,
+                                  confint = TRUE,
+                                  cutoff = 0.5,
+                                  bootstraps = 2500,
+                                  alpha = 0.05,
+                                  digits = 2) {
+  # Check if outcome is binary
+  unique_values <- unique(data[[outcome]])
+  if (!(length(unique_values) == 2 &&
+        all(unique_values %in% c(0, 1)))) {
+    stop("Outcome must be binary (containing only 0 and 1).")
+  }
 
+  pos_data <- data[data[[outcome]] == 1, ]
+  neg_data <- data[data[[outcome]] == 0, ]
+
+
+  # Need to work on conditional stats parity. Skipping for now
   if (!(is.null(group2) & is.null(condition))) {
     if (!group2 %in% names(data)) {
       stop("`group2` not found in data.")
     }
-    cond_stats_parity <- eval_cond_stats_parity(
-      data = data,
+
+    # check if the group2 is categorical or continuous
+    if (is.numeric(data[[group2]])) {
+      if (!grepl("^[<>=]", condition)) {
+        stop(
+          "Condition must be a character containing the sign of the condition
+           and the value to threshold the continuous variable
+           (e.g. '<50', '>50', '<=50', '>=50')."
+        )
+      } else {
+        subset_data <- subset(data, eval(parse(
+          text = paste0("data$", group2, condition)
+        )))
+      }
+    } else {
+      data[[group2]] <- as.factor(data[[group2]])
+      if (!condition %in% levels(data[[group2]])) {
+        stop("Condition must be a character of the levels to condition on.")
+      } else {
+        subset_data <- subset(data, data[[group2]] == condition)
+      }
+    }
+    cond_ppr <- get_ppr(
+      data = subset_data,
       outcome = outcome,
       group = group,
-      group2 = group2,
-      condition = condition,
       probs = probs,
-      confint = confint,
       cutoff = cutoff,
-      bootstraps = bootstraps,
-      alpha = alpha,
-      digits = digits,
-      message = FALSE
+      digits = digits
     )
+
+    cond_ppr_diff <- cond_ppr[[1]] - cond_ppr[[2]]
+    cond_ppr_ratio <- log(cond_ppr[[1]] / cond_ppr[[2]])
+
   }
 
-  stats_parity <- eval_stats_parity(
+  # Stats Parity
+  ppr <- get_ppr(
     data = data,
     outcome = outcome,
     group = group,
     probs = probs,
-    confint = confint,
     cutoff = cutoff,
-    bootstraps = bootstraps,
-    alpha = alpha,
-    digits = digits,
-    message = FALSE
-  )
-  eq_opp <- eval_eq_opp(
-    data = data,
-    outcome = outcome,
-    group = group,
-    probs = probs,
[…]
