### 2025-07-30 jianhuig/fairmetrics#25 @benyamindsmith
Hello @vankesteren,

1. As it stands, there are no p-values calculated in this package.
2. The CIs calculated are measures range associated with the difference/ratios of the model performance metrics.
3. In terms of action points, I want to confirm that you would like to see an emphasis on the the interpretation of the fairness metrics and their CIs should be based on the size of the effect and the uncertainty (measured in the form of a confidence interval) that yields evidence for a decision.
4. With point 3 raised, would you like to see the `message` parameter and its implementation removed from the `eval_*` functions? i.e.

https://github.com/jianhuig/fairmetrics/blob/cf5c1754cc4ef805329ed41bfdf54bdf8e466e89/R/eval_acc_parity.R#L124-L133
