+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: fairmetrics
-Title: Fairness Evaluation Metrics with Confidence Intervals
+Title: Fairness Evaluation Metrics with Confidence Intervals for Binary Sensitive Attributes
 Version: 1.0.4
 Authors@R: 
     c(
@@ -8,7 +8,7 @@ Authors@R:
     person("Benson", "Chou", email = "benson.chou@mail.utoronto.ca", role = c("aut"), comment = c(ORCID = "0009-0007-0265-033X")),
     person("Jessica", "Gronsbell", email = "j.gronsbell@utoronto.ca", role = c("aut"), comment = c(ORCID = "0000-0002-5360-5869"))
     )
-Description: A collection of functions for computing fairness metrics for machine learning and statistical models, including confidence intervals for each metric. The package supports the evaluation of group-level fairness criterion commonly used in fairness research, particularly in healthcare. It is based on the overview of fairness in machine learning written by Gao et al (2024) <doi:10.48550/arXiv.2406.09307>.
+Description: A collection of functions for computing fairness metrics for machine learning and statistical models, including confidence intervals for each metric. The package supports the evaluation of group-level fairness criterion commonly used in fairness research, particularly in healthcare for binary sensitive attributes. It is based on the overview of fairness in machine learning written by Gao et al (2024) <doi:10.48550/arXiv.2406.09307>.
 Imports:
   stats
 License: MIT + file LICENSE
