### 2025-08-04 ea30a1c
edited README, description, and paper
