+++ R/eval_acc_parity.R
@@ -1,12 +1,11 @@
 #' Examine Accuracy Parity of a Model
 #'
-#' This function assesses *Accuracy Parity*, a fairness criterion that evaluates whether
-#' the overall accuracy of a predictive model is consistent across different groups.
+#' This function assesses *Accuracy Parity*, a fairness criterion that evaluates whether the overall accuracy of a predictive model is consistent across two groups defined by a binary protected attribute.
 #'
 #' @param data Data frame containing the outcome, predicted outcome, and
-#' sensitive attribute
+#' binary protected attribute
 #' @param outcome Name of the outcome variable
-#' @param group Name of the sensitive attribute
+#' @param group Name of the binary protected attribute. Must consist of only two groups.
 #' @param probs Predicted probabilities
 #' @param cutoff Cutoff value for the predicted probabilities
 #' @param confint Logical indicating whether to calculate confidence intervals
@@ -46,7 +45,7 @@
 #' test_data$pred <- predict(rf_model, newdata = test_data, type = "prob")[, 2]
 #'
 #' # Fairness evaluation
-#' # We will use sex as the sensitive attribute and day_28_flg as the outcome.
+#' # We will use sex as the protected attribute and day_28_flg as the outcome.
 #' # We choose threshold = 0.41 so that the overall FPR is around 5%.
 #'
 #' # Evaluate Accuracy Parity
@@ -65,10 +64,14 @@
 eval_acc_parity <- function(data, outcome, group, probs, cutoff = 0.5, confint = TRUE,
                             alpha = 0.05, bootstraps = 2500,
                             digits = 2, message = TRUE) {
-  # Check if outcome is binary
+  # Check if outcome and groups are binary
   unique_values <- unique(data[[outcome]])
+  groups <- unique(data[[group]])
   if (!(length(unique_values) == 2 && all(unique_values %in% c(0, 1)))) {
-    stop("Outcome must be binary (containing only 0 and 1).")
+    stop("`outcome` must be binary (containing only 0 and 1).")
+  }
+  if (!(length(groups) == 2)) {
+    stop("`group` argument must only consist of two groups (i.e. `length(unique(data[[group]])) == 2`")
   }
 
   acc <- get_acc(
