### 2025-08-05 jianhuig/fairmetrics#24 @benyamindsmith
@vankesteren

1. I have now updated the [paper.md](https://github.com/jianhuig/fairmetrics/blob/main/utils/joss_paper/paper.md), [README](https://github.com/jianhuig/fairmetrics/blob/main/README.md) and [DESCRIPTION](https://github.com/jianhuig/fairmetrics/blob/main/DESCRIPTION) files to note that the package's usage is for assessing fairness among groups defined by a binary protected attributes (using the term _protected_ over sensitive).

2. Since the inference/estimation techniques used for more than two groups (e.g. see the [review paper](https://arxiv.org/abs/2406.09307), this is noted in the supplement). That being said, its possible to do "one-vs-all" comparisons with the existing functions- this can be added in the vingette (related to #23) and noted in the package documentation.

Let me know what you think.
