### 2025-08-06 7a13cb8
hiding kableExtra code
