### 2025-08-06 7b27b4f
not using fairmetrics::, just calling the function
