### 2025-08-06 5331537
rewriting vingette
