+++ vignettes/fairmetrics.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Tutorial: Assessing Model Fairness Across Binary Protected Attributes"
+title: "Assessing Model Fairness Across Binary Protected Attributes"
 authors:
   - name: "Jianhui Gao"
     orcid: "0000-0003-0915-1473"
@@ -32,11 +32,9 @@ knitr::opts_chunk$set(
 
 # Introduction
 
-We illustrate the usage of the `fairmetrics` package through a case study using a publicly available dataset of 1,776 ICU patients from the MIMIC-II clinical database, focusing on predicting 28-day mortality and evaluating disparities in model performance across sex.
+This vingette demonstrates how to obtain, report and interpret model fairness metrics for binary protected attributes with the `fairmetrics` package. We illustrate this through a case study based on a preprocessed version of the MIMIC II clinical database [1], which has been previously studied to explore the relationship between indwelling arterial catheters in hemodynamically stable patients and respiratory failure in relation to mortality outcomes [2]. The original, unprocessed dataset is publicly available through [PhysioNet](https://physionet.org/content/mimic2-iaccd/1.0/) [3]. A preprocessed version of this dataset is included in the `fairmetrics` package as the `mimic_preprocessed` and is used in this 
 
-The following packages are used for the analysis along with the `fairmetrics` package:
-
-```{r setup, message=FALSE, warning=FALSE}
+```{r setup, message=FALSE, warning=FALSE, include=FALSE}
 # Packages we are using for the analysis
 library(dplyr)
 library(corrplot)
@@ -44,153 +42,97 @@ library(randomForest)
 library(pROC)
 library(SpecsVerification)
 library(kableExtra)
-library(naniar)
 # Our package
 library(fairmetrics)
 ```
 
-The "Data Preprocessing" section discusses the dataset, the handling of missing data, model construction and standard predictive model evaluation through train-test splitting for binary classification. The "Fairness Evaluation" section shows how to evaluate the model's fairness toward binary protected attributes with the `fairmetrics` package.
+# Data Split and Model Construction
 
-```{=html}
-<!--
-Finally, the “Practical Considerations” section discusses the inherent trade-offs between different fairness definitions in clinical settings and highlights the challenges of enforcing fairness when outcome prevalence differs across groups. Details regarding the construction of 95% confidence intervals are provided in the appendix.
---->
-```
+In this setting, we construct a model that will predict 28-day mortality (`day_28_flg`). To do this, we split the dataset into a training and testing sets and fit a random forest model. The first 700 patients are used as the training set and the remaining patients are used as the testing set. 
 
-# Data Preprocessing
-
-The dataset used in this analysis is the MIMIC II clinical database [2], which has been previously studied to explore the relationship between indwelling arterial catheters in hemodynamically stable patients and respiratory failure in relation to mortality outcomes [3]. It includes 46 variables which cover demographics and clinical characteristics (including white blood cell count, heart rate during ICU stays and others) along with a 28-day mortality indicator (`day_28_flg`) for 1,776 patients. The data has been made publicly available by [PhysioNet](https://physionet.org/content/mimic2-iaccd/1.0/) [4] and is available in the `fairmetrics` package as the `mimic` dataset.
-
-## Handling Missing Data
-
[…]
