+++ vignettes/fairmetrics.Rmd
@@ -32,7 +32,7 @@ knitr::opts_chunk$set(
 
 # Introduction
 
-This vingette demonstrates how to obtain, report and interpret model fairness metrics for binary protected attributes with the `fairmetrics` package. We illustrate this through a case study based on a preprocessed version of the MIMIC II clinical database [1], which has been previously studied to explore the relationship between indwelling arterial catheters in hemodynamically stable patients and respiratory failure in relation to mortality outcomes [2]. The original, unprocessed dataset is publicly available through [PhysioNet](https://physionet.org/content/mimic2-iaccd/1.0/) [3]. A preprocessed version of this dataset is included in the `fairmetrics` package as the `mimic_preprocessed` and is used in this 
+This vingette demonstrates how to obtain, report and interpret model fairness metrics for binary protected attributes with the `fairmetrics` package. We illustrate this through a case study based on a preprocessed version of the MIMIC II clinical database [1], which has been previously studied to explore the relationship between indwelling arterial catheters in hemodynamically stable patients and respiratory failure in relation to mortality outcomes [2]. The original, unprocessed dataset is publicly available through [PhysioNet](https://physionet.org/content/mimic2-iaccd/1.0/) [3]. A preprocessed version of this dataset is included in the `fairmetrics` package as the `mimic_preprocessed` and is used in this vingette.
 
 ```{r setup, message=FALSE, warning=FALSE, include=FALSE}
 # Packages we are using for the analysis
@@ -48,9 +48,7 @@ library(fairmetrics)
 
 # Data Split and Model Construction
 
-In this setting, we construct a model that will predict 28-day mortality (`day_28_flg`). To do this, we split the dataset into a training and testing sets and fit a random forest model. The first 700 patients are used as the training set and the remaining patients are used as the testing set. 
-
-After the model is fit, it is used to predict 28-day mortality. The predicted probabilities are saved as new column in the testing data and are used to assess model fairness. 
+In this setting, we construct a model that will predict 28-day mortality (`day_28_flg`). To do this, we split the dataset into a training and testing sets and fit a random forest model. The first 700 patients are used as the training set and the remaining patients are used as the testing set. After the model is fit, it is used to predict 28-day mortality. The predicted probabilities are saved as new column in the testing data and are used to assess model fairness. 
 
 ```{r}
 # Load required libraries
@@ -58,7 +56,7 @@ library(dplyr)
 library(pROC)
 library(fairmetrics)
 # Set seed for reproducibility
-set.seed(123)
+set.seed(1)
 # Load the MIMIC-II preprocessed data
 data("mimic_preprocessed")
 # Use 700 labels to train the mimic
@@ -77,7 +75,7 @@ test_data$pred <- predict(rf_model, newdata = test_data, type = "prob")[, 2]
 
 # Fairness Evaluation
 
-The `fairmetrics` package is used to assess model fairness accross binary protected attributes. This means that the unique values in the protected attribute column need to be exactly two. To evaluate fairness of the random forest model which we fit, we examine patient gender as the binary protected attribute.
+The `fairmetrics` package is used to assess model fairness across binary protected attributes. This means that the unique values in the protected attribute column need to be exactly two. To evaluate fairness of the random forest model which we fit, we examine patient gender as the binary protected attribute.
 
 ```{r}
 # Recode gender variable explicitly for readability: 
@@ -100,7 +98,7 @@ eval_stats_parity(
 
[…]
