+++ utils/joss_paper/paper.Rmd
@@ -83,21 +83,21 @@ get_fairness_metrics(
 )
 
 $performance
-                                    Metric Group1 Group2
-1                 Positive Prediction Rate   0.17   0.08
-2     Conditional Positive Prediction Rate   0.34   0.21
-3                      False Negative Rate   0.38   0.62
-4                      False Positive Rate   0.08   0.03
-5            Avg. Predicted Positive Prob.   0.46   0.37
-6            Avg. Predicted Negative Prob.   0.15   0.10
-7                Positive Predictive Value   0.62   0.66
-8                Negative Predictive Value   0.92   0.90
-9                              Brier Score   0.09   0.08
-10                                Accuracy   0.87   0.88
-11 (False Negative)/(False Positive) Ratio   1.03   3.24
+                                    Metric GroupFemale GroupMale
+1                 Positive Prediction Rate        0.17      0.08
+2     Conditional Positive Prediction Rate        0.34      0.21
+3                      False Negative Rate        0.38      0.62
+4                      False Positive Rate        0.08      0.03
+5            Avg. Predicted Positive Prob.        0.46      0.37
+6            Avg. Predicted Negative Prob.        0.15      0.10
+7                Positive Predictive Value        0.62      0.66
+8                Negative Predictive Value        0.92      0.90
+9                              Brier Score        0.09      0.08
+10                                Accuracy        0.87      0.88
+11 (False Negative)/(False Positive) Ratio        1.03      3.24
 
 $fairness
-                           Metric Difference        Diff_CI Ratio     Ratio_CI
+                           Metric Difference    95% Diff CI Ratio 95% Ratio CI
 1              Statistical Parity       0.09   [0.05, 0.13]  2.12 [1.48, 3.05]
 2  Conditional Statistical Parity       0.13   [0.05, 0.21]  1.62 [1.17, 2.24]
 3               Equal Opportunity      -0.24 [-0.39, -0.09]  0.61 [0.44, 0.86]
