+++ R/eval_neg_pred_parity.R
@@ -120,9 +120,9 @@ eval_neg_pred_parity <- function(data, outcome, group, probs, cutoff = 0.5, conf
       paste0("Group", sort(unique(data[[group]]))[1]),
       paste0("Group", sort(unique(data[[group]]))[2]),
       "Difference",
-      "95% Diff CI",
+      paste0((1-alpha)*100, "% Diff CI"),
       "Ratio",
-      "95% Ratio CI"
+      paste0((1-alpha)*100, "% Ratio CI")
     )
 
 
