+++ R/eval_neg_class_bal.R
@@ -118,9 +118,9 @@ eval_neg_class_bal <- function(data, outcome, group, probs, confint = TRUE,
       paste0("Group", sort(unique(data[[group]]))[1]),
       paste0("Group", sort(unique(data[[group]]))[2]),
       "Difference",
-      "95% Diff CI",
+      paste0((1-alpha)*100, "% Diff CI"),
       "Ratio",
-      "95% Ratio CI"
+      paste0((1-alpha)*100, "% Ratio CI")
     )
 
     if (message) {
