+++ R/eval_acc_parity.R
@@ -118,9 +118,9 @@ eval_acc_parity <- function(data, outcome, group, probs, cutoff = 0.5, confint =
       paste0("Group", sort(unique(data[[group]]))[1]),
       paste0("Group", sort(unique(data[[group]]))[2]),
       "Difference",
-      "95% Diff CI",
+      paste0((1-alpha)*100, "% Diff CI"),
       "Ratio",
-      "95% Ratio CI"
+      paste0((1-alpha)*100, "% Ratio CI")
     )
 
 
