### 2025-08-11 jianhuig/fairmetrics#25 @vankesteren
> 1. As it stands, there are no p-values calculated in this package.

but the package does do statistical inference, with the same pitfall: it suggests null hypothesis significance testing via the CI, which is equivalent because you produce 95% CIs and produce binary "decisions".

Your message in point 4 is correct and nicely phrased, and I don't think you should remove it. I just think that additionally you should discuss / address the issue I raise with clear visibility to package users to prevent them from misinterpreting null findings.
