### 2025-08-11 jianhuig/fairmetrics#23 @vankesteren
Nice start, but I found 2 errors in the first chunk when I look [here](https://jianhuig.github.io/fairmetrics/articles/fairmetrics.html) -- this is the newest version right? If so, please check the whole document again to make sure it is reproducible and correct.

Error after chunk 1:
```
Error in randomForest(factor(day_28_flg) ~ ., data = train_data, ntree = 1000) :
  could not find function "randomForest"
```

(you need `library(randomforest)`)

Also here:
```R
# Load the MIMIC-II preprocessed data
data("mimic_preprocessed")
# Use 700 labels to train the mimic
train_data <- mimic_preprocessed %>%
  filter(row_number() <= 700)

# Test the model on the remaining data
test_data <- mimic %>%
  filter(row_number() > 700)
```

you use the `mimic_preprocessed` for the train data, but you use `mimic` (the not preprocessed one) for the test data; please homogenize. Also, the `data()` call is not needed I think.
