### 2025-08-11 jianhuig/fairmetrics#23 @vankesteren
oh and the in-text reference links don't seem to work correctly (the numbers between square brackets do not hyperlink to the references, wasn't sure if this was intentional)
