### 2025-08-11 #8497 @vankesteren
I have made some updates to the discussions in the open issues. One is pretty much fixed now (https://github.com/jianhuig/fairmetrics/issues/24), but I want to be clear that I think the package does need some improvement on the user-friendliness and robustness before I can accept it.

My feeling is that this package is quite light in its contribution (there being several other packages that do similar things), and some of the features in this package (like conditioning) are IMO better solved by other means (see my comments in https://github.com/jianhuig/fairmetrics/issues/27). If that is excluded, the main contributions are:
1) compilation of the metrics
2) bootstrapped confidence intervals

And specifically only for binary protected attributes (so not for e.g., migration background; https://github.com/jianhuig/fairmetrics/issues/24)

From my point of view, this _can_ be enough of a contribution if the package is really robustly built (e.g., with a fully tested API; see my latest comment here: https://github.com/jianhuig/fairmetrics/issues/27), supported by great documentation (see requests here: https://github.com/jianhuig/fairmetrics/issues/23), and guides the user in computation, presentation, and interpretation of the metrics (https://github.com/jianhuig/fairmetrics/issues/27 and https://github.com/jianhuig/fairmetrics/issues/25).

Let me note that I really like the package idea and I want it to succeed, be supported, and used widely for the foreseeable future; please forgive me if I come across as overly critical! I am happy to have @StevenGolovkine or @Nikoleta-v3 add their points of view in these issues if that helps.
