+++ R/get_fairness_metrics.R
@@ -1,6 +1,6 @@
 #' Compute Fairness Metrics for Binary Classification
 #'
-#' Computes a comprehensive set of fairness metrics for binary classification models, disaggregated by a binary protected attribute (e.g., race, gender). Optionally, conditional fairness can be evaluated using a second attribute and a specified condition. The function also computes corresponding performance metrics used in the fairness calculations.
+#' Computes a comprehensive set of fairness metrics for binary classification models, disaggregated by a binary protected attribute. The function also computes corresponding performance metrics used in the fairness calculations.
 #'
 #' The results are returned as a list of two data frames:
 #' - `performance`: Contains performance metrics (e.g., TPR, FPR, PPV) by group.
@@ -8,8 +8,6 @@
 #'
 #' ### Fairness Metrics Included:
 #' - **Statistical Parity**: Difference in positive prediction rates across groups.
-#' - **Conditional Statistical Parity** *(if group2 and condition are specified)*:
-#'   Parity conditioned on a second group and value.
 #' - **Equal Opportunity**: Difference in true positive rates (TPR) across groups.
 #' - **Predictive Equality**: Difference in false positive rates (FPR) across groups.
 #' - **Balance for Positive Class**: Checks whether the predicted probability distributions for
@@ -24,12 +22,6 @@
 #' @param data A data frame containing the outcome, group, and predicted probabilities.
 #' @param outcome The name of the column containing the true binary outcome.
 #' @param group The name of the column representing the binary protected attribute (e.g., race, gender).
-#' @param group2 Define if conditional statistical parity is desired. Name of a secondary group variable used for conditional fairness analysis.
-#' @param condition Define if conditional statistical parity is desired. If the conditional group is categorical, the condition
-#' supplied must be a character of the levels to condition on. If the conditional
-#' group is continuous, the conditions supplied must be a character containing
-#' the sign of the condition and the value to threshold the continuous variable
-#' (e.g. "<50", ">50", "<=50", ">=50").
 #' @param probs The name of the column with predicted probabilities.
 #' @param confint Logical indicating whether to calculate confidence intervals.
 #' @param cutoff Numeric threshold for classification. Default is 0.5.
@@ -71,8 +63,6 @@
 #'  data = test_data,
 #'  outcome = "day_28_flg",
 #'  group = "gender",
-#'  group2 = "age",
-#'  condition = ">=60",
 #'  probs = "pred",
 #'  confint = TRUE,
 #'  cutoff = 0.41,
@@ -83,71 +73,30 @@
 #'
 #' @export
 get_fairness_metrics <- function(data,
-                                  outcome,
-                                  group,
-                                  group2 = NULL,
-                                  condition = NULL,
-                                  probs,
-                                  confint = TRUE,
-                                  cutoff = 0.5,
-                                  bootstraps = 2500,
-                                  alpha = 0.05,
-                                  digits = 2) {
+                                 outcome,
[…]
