### 2025-08-11 62be4fb
updating vingette and  JOSS article
