+++ utils/joss_paper/paper.md
@@ -67,9 +67,7 @@ models across binary protected attributes. The package supports the
 evaluation of metrics belonging to three major group fairness criteria:
 
 -   **Independence:** Statistical Parity (compares the overall rate of
-    positive predictions between groups), Conditional Statistical Parity
-    (compares the rate of positive predictions between groups within a
-    specific subgroup).
+    positive predictions between groups).
 
 -   **Separation:** Equal Opportunity (compares false negative rates
     between groups), Predictive Equality (compares false positive rates
@@ -113,9 +111,8 @@ Raffa 2016; J. D. Raffa et al. 2016).
 While the choice of fairness metric used is context dependent, we show
 all criteria available with the `get_fairness_metrics()` function for
 illustrative purposes. In this example, we evaluate the model’s fairness
-with respect to the binary protected attribute `gender`. For conditional
-statistical parity, we condition on age greater than 60 years. The model
-is trained on a subset of the data and the predictions are made and
+with respect to the binary protected attribute `gender`. The model is
+trained on a subset of the data and the predictions are made and
 evaluated on a test set. A statistically significant difference across
 groups at a given level of significance is indicated when the confidence
 interval for a difference-based metric does not include zero or when the
@@ -130,8 +127,6 @@ interval for a ratio-based metric does not include one.
      data = test_data,
      outcome = "day_28_flg",
      group = "gender",
-     group2 = "age",
-     condition = ">=60",
      probs = "pred",
      cutoff = 0.41, 
      alpha = 0.05
@@ -140,30 +135,28 @@ interval for a ratio-based metric does not include one.
     $performance
                                         Metric GroupFemale GroupMale
     1                 Positive Prediction Rate        0.17      0.08
-    2     Conditional Positive Prediction Rate        0.34      0.21
-    3                      False Negative Rate        0.38      0.62
-    4                      False Positive Rate        0.08      0.03
-    5            Avg. Predicted Positive Prob.        0.46      0.37
-    6            Avg. Predicted Negative Prob.        0.15      0.10
-    7                Positive Predictive Value        0.62      0.66
-    8                Negative Predictive Value        0.92      0.90
-    9                              Brier Score        0.09      0.08
-    10                                Accuracy        0.87      0.88
-    11 (False Negative)/(False Positive) Ratio        1.03      3.24
+    2                      False Negative Rate        0.38      0.62
+    3                      False Positive Rate        0.08      0.03
+    4            Avg. Predicted Positive Prob.        0.46      0.37
+    5            Avg. Predicted Negative Prob.        0.15      0.10
+    6                Positive Predictive Value        0.62      0.66
+    7                Negative Predictive Value        0.92      0.90
+    8                              Brier Score        0.09      0.08
+    9                                Accuracy        0.87      0.88
+    10 (False Negative)/(False Positive) Ratio        1.03      3.24
 
     $fairness
                                Metric Difference    95% Diff CI Ratio 95% Ratio CI
     1              Statistical Parity       0.09   [0.05, 0.13]  2.12 [1.48, 3.05]
-    2  Conditional Statistical Parity       0.13   [0.05, 0.21]  1.62 [1.17, 2.24]
[…]
