+++ utils/joss_paper/paper.Rmd
@@ -46,7 +46,7 @@ ML models are increasingly used in high-stakes domains such as criminal justice,
 
 `fairmetrics` is designed to evaluate fairness of binary classification models across binary protected attributes. The package supports the evaluation of metrics belonging to three major group fairness criteria: 
 
-- **Independence:** Statistical Parity (compares the overall rate of positive predictions between groups), Conditional Statistical Parity (compares the rate of positive predictions between groups within a specific subgroup).
+- **Independence:** Statistical Parity (compares the overall rate of positive predictions between groups).
 
 - **Separation:** Equal Opportunity (compares false negative rates between groups), Predictive Equality (compares false positive rates between groups), Balance for Positive Class (compares the average predicted probabilities among individuals whose true outcome is positive across groups), and Balance for Negative Class (compares the average predicted probabilities among individuals whose true outcome is negative across groups).
   
@@ -62,7 +62,7 @@ The input required to evaluate model fairness with the `fairmetrics` package is 
 
 A simple example of how to use the `fairmetrics` package is illustrated below. The example makes use of the `mimic_preprocessed` dataset, a pre-processed version of the Indwelling Arterial Catheter (IAC) Clinical dataset, from the MIMIC-II clinical database [@goldberger2000physiobank; @raffa2016clinical; @raffa2016data]. 
 
-While the choice of fairness metric used is context dependent, we show all criteria available with the `get_fairness_metrics()` function for illustrative purposes. In this example, we evaluate the model's fairness with respect to the binary protected attribute `gender`. For conditional statistical parity, we condition on age greater than 60 years. The model is trained on a subset of the data and the predictions are made and evaluated on a test set.  A statistically significant difference across groups at a given level of significance is indicated when the confidence interval for a difference-based metric does not include zero or when the interval for a ratio-based metric does not include one.
+While the choice of fairness metric used is context dependent, we show all criteria available with the `get_fairness_metrics()` function for illustrative purposes. In this example, we evaluate the model's fairness with respect to the binary protected attribute `gender`. The model is trained on a subset of the data and the predictions are made and evaluated on a test set.  A statistically significant difference across groups at a given level of significance is indicated when the confidence interval for a difference-based metric does not include zero or when the interval for a ratio-based metric does not include one.
 
 
 ```r
@@ -75,8 +75,6 @@ get_fairness_metrics(
  data = test_data,
  outcome = "day_28_flg",
  group = "gender",
- group2 = "age",
- condition = ">=60",
  probs = "pred",
  cutoff = 0.41, 
  alpha = 0.05
@@ -85,30 +83,28 @@ get_fairness_metrics(
 $performance
                                     Metric GroupFemale GroupMale
 1                 Positive Prediction Rate        0.17      0.08
-2     Conditional Positive Prediction Rate        0.34      0.21
-3                      False Negative Rate        0.38      0.62
-4                      False Positive Rate        0.08      0.03
-5            Avg. Predicted Positive Prob.        0.46      0.37
-6            Avg. Predicted Negative Prob.        0.15      0.10
-7                Positive Predictive Value        0.62      0.66
-8                Negative Predictive Value        0.92      0.90
[…]
