### 2025-08-11 jianhuig/fairmetrics#27 @benyamindsmith
@vankesteren

1. The model performance metrics match 1-to-1 with the model fairness metrics (i.e. PPR and Statistical Parity, FNR and Equal Opportunity etc.) we can mention this in the documentation or make it a single dataframe. The only caveat would be that the single dataframe may not be able to fit across the page of the JOSS manuscript (I'm unsure if any further typesetting is preformed by the JOSS team before publication beyond just using the bot or not).

2. There are tests for get_fairness_metrics which should be sufficient now since conditional statistical parity is not included in the package (see #21 )

Let me know what you think.
