+++ tests/testthat/test-get_fairness_metrics-no-errors.R
@@ -5,8 +5,6 @@ test_that('get_fairness_metrics CI test', {
       data = test_data,
       outcome = "day_28_flg",
       group = "gender",
-      group2 = "age",
-      condition = ">=60",
       probs = "pred",
       confint = TRUE,
       cutoff = 0.41,
@@ -25,8 +23,6 @@ test_that('get_fairness_metrics no CI test', {
       data = test_data,
       outcome = "day_28_flg",
       group = "gender",
-      group2 = "age",
-      condition = ">=60",
       probs = "pred",
       confint = FALSE,
       cutoff = 0.41,
