### 2025-08-11 ffc94e8
editing vingette
