### 2025-08-11 jianhuig/fairmetrics#23 @benyamindsmith
@vankesteren

Good catch, thank you for noticing this error. I just read through the vingette again and it looks good on my end.

The summary of significant changes are:

1. The inclusion of `library(randomForest)` in code chunk 1.
2. Removal of the `data()`call for `mimic_preprocessed`
3. Using `mimic_preprocessed` for both train and test datasets

As far as the inline reference links are concerned  - them being non-linking is intentional as I haven't been able to use a `.bib` file for reference management.
