### 2025-08-21 jianhuig/fairmetrics#23 @vankesteren
One last comment and 1 typo:

Now, you don't show CIs at all in the vignette, even though this is one of the main features of your package! I think this should be in the get_fairness_metrics part.

Typo in one of the last paragraphs:

> While fairmetrics focuses only of assessing fairness ...

should be
> While fairmetrics focuses only __on__ assessing fairness ...
