### 2025-08-21 8ae8505
updating get_fairness_metrics to return a dataframe
