+++ vignettes/fairmetrics.Rmd
@@ -97,7 +97,7 @@ get_fairness_metrics(
  )
 ```
 
-While `fairmetrics` focuses only of assessing fairness of models across binary protected attributes, it is possible to work with protected attributes which consist of more than two groups by using "one-vs-all" comparisons and a little bit of data wrangling to create the appropriate columns. 
+While `fairmetrics` focuses only on assessing fairness of models across binary protected attributes, it is possible to work with protected attributes which consist of more than two groups by using "one-vs-all" comparisons and a little bit of data wrangling to create the appropriate columns. 
 
 
 # Appendix: Confidence Interval Construction
