+++ vignettes/fairmetrics.Rmd
@@ -28,6 +28,7 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
+set.seed(123)
 ```
 
 # Introduction
@@ -85,7 +86,7 @@ eval_stats_parity(
 
 The dataframe returned gives the positive prediction rate between the groups defined by the binary protected attribute (`GroupFemale` and `GroupMale` in this case), the difference and ratios between the groups and the bootstrap calculated confidence intervals for the estimated difference and ratios. For inference, it can be considered that confidence interval which contains 0 in its range for difference or 1 for ratio as insignificant.
 
-The above syntax can be extended to the other `eval_*` functions as well. Additionally, if one wishes to calculate various fairness metrics for the model simultaneously, we pass our test data with its predicted results into the `get_fairness_metrics` function.
+The above syntax can be extended to the other `eval_*` functions as well. Additionally, if one wishes to calculate various fairness metrics (and their confidence intervals) for the model simultaneously, we pass our test data with its predicted results into the `get_fairness_metrics` function.
 
 ```{r}
 get_fairness_metrics(
@@ -96,6 +97,12 @@ get_fairness_metrics(
   cutoff = 0.41
  )
 ```
+
+With `get_fairness_metrics`, it is possible to see through the bootstrapped confidence intervals which fairness critereon are violated and which are not. From the above output, we observe that Statistical Parity, Equal Opportunity, Predictive Equality, Balance for Positive Class and Balance for Negative Class possess some evidence of being violated, while Positive Predictive Parity, Negative Predictive Parity, Brier Score Parity and Overall Accuracy Parity do not. Treatment equality remains ambiguous based on the measure. If  difference is considered, the 95%  bootstrapped confidence interval crosses 0 - indicating that there is not evidence indicating that the fairness criterion is violated, while the confidence interval for the ratio crosses 1 - indicating that there is. 
+
+>[!WARNING]
+> Statistical inference from bootstrapped confidence intervals should be interpreted with caution. A confidence interval crossing 0 (for differences) or 1 (for ratios) means the evidence is inconclusive rather than proving the absence of unfairness. Apparent violations may reflect sampling variability rather than systematic bias. Always complement these results with domain knowledge, sensitivity analyses, and additional fairness diagnostics before drawing strong conclusions.  
+
 
 While `fairmetrics` focuses only on assessing fairness of models across binary protected attributes, it is possible to work with protected attributes which consist of more than two groups by using "one-vs-all" comparisons and a little bit of data wrangling to create the appropriate columns. 
 
