+++ vignettes/fairmetrics.Rmd
@@ -98,14 +98,42 @@ get_fairness_metrics(
  )
 ```
 
-With `get_fairness_metrics`, it is possible to see through the bootstrapped confidence intervals which fairness critereon are violated and which are not. From the above output, we observe that Statistical Parity, Equal Opportunity, Predictive Equality, Balance for Positive Class and Balance for Negative Class possess some evidence of being violated, while Positive Predictive Parity, Negative Predictive Parity, Brier Score Parity and Overall Accuracy Parity do not. Treatment equality remains ambiguous based on the measure. If  difference is considered, the 95%  bootstrapped confidence interval crosses 0 - indicating that there is not evidence indicating that the fairness criterion is violated, while the confidence interval for the ratio crosses 1 - indicating that there is. 
+With the `get_fairness_metrics` function, it's possible to examine the bootstrapped confidence intervals to determine which fairness criterion show evidence of violation. From the above output, Statistical Parity, Equal Opportunity, Predictive Equality, Balance for Positive Class and Balance for Negative Class show  evidence of being violated, while Positive Predictive Parity, Negative Predictive Parity, Brier Score Parity and Overall Accuracy Parity do not. Treatment Equality remains ambiguous: when assessed by difference, the 95%  bootstrapped confidence interval crosses 0 - indicating that there is no evidence indicating that the fairness criterion is violated, but when assessed by ratio, the interval crosses 1 - which indicates that there is. 
 
 > __NOTE:__
 >
-> Statistical inference from bootstrapped confidence intervals should be interpreted with caution. A confidence interval crossing 0 (for differences) or 1 (for ratios) means the evidence is inconclusive rather than proving the absence of unfairness. Apparent violations may reflect sampling variability rather than systematic bias. Always complement these results with domain knowledge, sensitivity analyses, and additional fairness diagnostics before drawing strong conclusions.  
+> Statistical inference from bootstrapped confidence intervals should be interpreted with caution. A confidence interval crossing 0 (for differences) or 1 (for ratios) means the evidence is inconclusive rather than proving absence of unfairness. Apparent violations may reflect sampling variability rather than systematic bias. Always complement these results with domain knowledge, sensitivity analyses, and additional fairness diagnostics before drawing strong conclusions about a specific fairness assesment.
 
+While `fairmetrics` focuses only on assessing fairness of models across binary protected attributes, it is possible to work with protected attributes which consist of more than two groups by using "one-vs-all" comparisons and a little bit of data wrangling to create the appropriate columns.
 
-While `fairmetrics` focuses only on assessing fairness of models across binary protected attributes, it is possible to work with protected attributes which consist of more than two groups by using "one-vs-all" comparisons and a little bit of data wrangling to create the appropriate columns. 
+# Conditional Fairness Evaluation
+
+To evaluate a model's fairness for a specific subgroup, simply subset the data. For example, to evaluate the conditional statistical parity among subjects aged 60 and older using the previously fitted random forest model, subsetting the test data and and passing it through the `eval_stats_parity` is all thats required. 
+
+```{r}
+eval_stats_parity(
+  data = subset(test_data, age >= 60), 
+  outcome = "day_28_flg",
+  group = "gender",
+  probs = "pred",
+  cutoff = 0.41,
+  message = TRUE
+)
+```
+
+From the bootstrapped confidence intervals for the difference and the ratio, we note that there is evidence that statistical parity is violated based on the difference and ratio confidence intervals between the male and female subjects aged 60 and over. A similar example can be constructed with `get_fairness_metrics` for a more comprehensive view of the conditional fairness of the model. 
+
+```{r}
+get_fairness_metrics(
[…]
