### 2025-08-22 jianhuig/fairmetrics#23 @benyamindsmith
@vankesteren

1. I edited the vignette to include commentary on the confidence intervals (and also addressed #21 with discussion about conditional fairness and #25 with a pull quote noting the limitations of inference from the bootstrapped CIs).

2. Fixed the typo.

Check out [0331e9f](https://github.com/jianhuig/fairmetrics/commit/0331e9f0844b512b4c547408b810a076bec8b349) (for reference).

I'll also include the section of text here.

https://github.com/jianhuig/fairmetrics/blob/0331e9f0844b512b4c547408b810a076bec8b349/vignettes/fairmetrics.Rmd#L101-L136
