### 2025-08-22 48427f1
documentation and typo fix
