### 2025-08-22 jianhuig/fairmetrics#25 @benyamindsmith
@vankesteren  Thanks for the feedback here.

1. I added this note to the `get_fairness_metrics` documentation:

https://github.com/jianhuig/fairmetrics/blob/a2f80b9a44aa520a1c095699d37fbe47baeaf568/R/get_fairness_metrics.R#L22

2. This note is echoed in the vignette.
    https://github.com/jianhuig/fairmetrics/blob/a2f80b9a44aa520a1c095699d37fbe47baeaf568/vignettes/fairmetrics.Rmd#L103-L105
   I wanted to do a warning callout, but I but I was unable to do it (it appears that this is an outstanding issue, see: https://github.com/r-lib/pkgdown/issues/2701)

For actual history on how this was written, the note was first written in the vignette and was copied over to the `get_fairness_metrics` documentation.

Let me know what you think
