### 2025-08-22 6b7f0c1
adding RCMD check with github actions
