### 2025-09-03 #8497 @vankesteren
Yeah, semester had started again which means loads of teaching/supervision duties, but the issues are going well and we're almost done. I'll see if I can give the review a final push this week 👍
