### 2025-09-07 #8497 @benyamindsmith
Thank you reviewers @StevenGolovkine  and @vankesteren for your thorough feedback on the `fairmetrics` package and seeing its merit as a valuable contribution!

@Nikoleta-v3 keep me posted on next steps.
