### 2025-09-07 b81f4b9
small fix 'Department of Statistical Sciences'
