+++ utils/joss_paper/paper.md
@@ -19,7 +19,7 @@ authors:
     orcid: 0000-0002-5360-5869
     affiliation: 1
 affiliations:
-  - name: "Department of Statistical Science, University of Toronto"
+  - name: "Department of Statistical Sciences, University of Toronto"
     index: 1
 date: "2025-06-10"
 bibliography: paper.bib
