+++ utils/joss_paper/paper.tex
@@ -274,7 +274,7 @@ evaluation. \textit{Journal of Open Source Software}, (), . \href{https://doi.or
           \author[1]{Benson Chou}
           \author[1]{Jessica Gronsbell}
     
-      \affil[1]{Department of Statistical Science, University of
+      \affil[1]{Department of Statistical Sciences, University of
 Toronto}
   \date{\vspace{-5ex}}
 
