### 2025-09-10 813a963
fixing style in bib entries
