+++ utils/joss_paper/paper.bib
@@ -33,7 +33,7 @@ series = {FAccT '22}
   year = {2016},
   urldate = {2025-05-29},
   abstract = {There's software used across the country to predict future criminals. And it's biased against blacks.},
-  howpublished = {https://www.propublica.org/article/machine-bias-risk-assessments-in-criminal-sentencing},
+  howpublished = {\url{https://www.propublica.org/article/machine-bias-risk-assessments-in-criminal-sentencing}},
   langid = {english}
 }
 
@@ -172,7 +172,7 @@ series = {ITCS '12}
 @article{Castelnovo_Crupi_Greco_Regoli_Penco_Cosentini_2022, title={A clarification of the nuances in the fairness metrics landscape}, volume={12}, url={https://www.nature.com/articles/s41598-022-07939-1}, DOI={10.1038/s41598-022-07939-1}, number={1}, journal={Scientific Reports}, author={Castelnovo, Alessandro and Crupi, Riccardo and Greco, Greta and Regoli, Daniele and Penco, Ilaria Giuseppina and Cosentini, Andrea Claudio}, year={2022}, month=mar }
 
  @Manual{fairness_package,
-    title = {fairness: Algorithmic Fairness Metrics},
+    title = {{f}airness: Algorithmic Fairness Metrics},
     author = {Nikita Kozodoi and Tibor {V. Varga}},
     year = {2021},
     note = {R package version 1.2.1},
@@ -184,7 +184,6 @@ series = {ITCS '12}
   author       = {Raffa, Jesse},
   title        = {Clinical data from the MIMIC-II database for a case study on indwelling arterial catheters (version 1.0)},
   year         = {2016},
-  howpublished = {\url{https://doi.org/10.13026/C2NC7F}},
   note         = {PhysioNet},
   doi          = {10.13026/C2NC7F}
 }
@@ -212,7 +211,7 @@ series = {ITCS '12}
 }
 
 @article{wisniewski2022fairmodels,
-  title     = {fairmodels: a Flexible Tool for Bias Detection, Visualization, and Mitigation in Binary Classification Models},
+  title     = {{f}airmodels: a Flexible Tool for Bias Detection, Visualization, and Mitigation in Binary Classification Models},
   author    = {Jakub Wiśniewski and Przemysław Biecek},
   journal   = {The R Journal},
   year      = {2022},
