### 2025-09-10 #8497 @Nikoleta-v3
Hey @benyamindsmith 👋🏻 I opened a pull request to tidy up the bibliography file (https://github.com/jianhuig/fairmetrics/pull/29). I’m not sure if I broke anything 😅 could you review and merge if it looks good? Then we can compile the paper.

One thing: there’s a typo in the paper/code. In the data frame returned by `get_fairness_metrics`, the column label should be *Fairness Assessment*, not *Fairness Assesment*. Could you fix this please? 😄
