+++ utils/joss_paper/paper.md
@@ -24,114 +24,213 @@ affiliations:
 date: "2025-06-10"
 bibliography: paper.bib
 output:
-  rticles::joss_article:
-    keep_tex: true
   md_document:
      #keep_tex: true
      preserve_yaml: TRUE
      variant: "markdown_strict"
+  rticles::joss_article:
+    keep_tex: true
 journal: JOSS
 ---
 
 # Summary
 
-Fairness is a growing area of machine learning (ML) that focuses on ensuring that models do not produce systematically biased outcomes across groups defined by protected attributes, such as race, gender, or age. The `fairmetrics` R package provides a user-friendly framework for rigorously evaluating group-based fairness criteria, including independence (e.g., statistical parity), separation (e.g., equalized odds), and sufficiency (e.g., predictive parity) for binary protected attributes. The package provides both point and interval estimates for a variety of commonly used criteria. `fairmetrics` also includes an example dataset derived from the Medical Information Mart for Intensive Care, version II (MIMIC-II) database [@goldberger2000physiobank; @raffa2016clinical] to demonstrate its use. 
-
+Fairness is a growing area of machine learning (ML) that focuses on
+ensuring that models do not produce systematically biased outcomes
+across groups defined by protected attributes, such as race, gender, or
+age. The `fairmetrics` R package provides a user-friendly framework for
+rigorously evaluating group-based fairness criteria, including
+independence (e.g., statistical parity), separation (e.g., equalized
+odds), and sufficiency (e.g., predictive parity) for binary protected
+attributes. The package provides both point and interval estimates for a
+variety of commonly used criteria. `fairmetrics` also includes an
+example dataset derived from the Medical Information Mart for Intensive
+Care, version II (MIMIC-II) database (Goldberger et al. 2000; J. Raffa
+2016) to demonstrate its use.
 
 # Statement of Need
 
-ML models are increasingly used in high-stakes domains such as criminal justice, healthcare, finance, employment, and education [@mehrabi_survey_21; @mattuMachineBias; @Gao_Chou_McCaw_Thurston_Varghese_Hong_Gronsbell_2024]. Existing fairness evaluation software report point estimates and/or visualizations, without any measures of uncertainty. This limits users' ability to determine whether observed disparities are statistically significant. `fairmetrics` addresses this limitation by including confidence intervals for both difference and ratio based fairness metrics to enable more robust and statistically grounded fairness assessments. 
+ML models are increasingly used in high-stakes domains such as criminal
+justice, healthcare, finance, employment, and education (Mehrabi et al.
+2021; Mattu 2016; Gao et al. 2024). Existing fairness evaluation
+software report point estimates and/or visualizations, without any
+measures of uncertainty. This limits users’ ability to determine whether
+observed disparities are statistically significant. `fairmetrics`
+addresses this limitation by including confidence intervals for both
+difference and ratio based fairness metrics to enable more robust and
+statistically grounded fairness assessments.
 
 # Fairness Criteria
 
-`fairmetrics` is designed to evaluate fairness of binary classification models across binary protected attributes. The package supports the evaluation of metrics belonging to three major group fairness criteria: 
+`fairmetrics` is designed to evaluate fairness of binary classification
+models across binary protected attributes. The package supports the
[…]
