+++ utils/joss_paper/paper.Rmd
@@ -24,12 +24,12 @@ affiliations:
 date: "2025-06-10"
 bibliography: paper.bib
 output:
-  rticles::joss_article:
-    keep_tex: true
   md_document:
      #keep_tex: true
      preserve_yaml: TRUE
      variant: "markdown_strict"
+  rticles::joss_article:
+    keep_tex: true
 journal: JOSS
 ---
 
@@ -80,7 +80,7 @@ get_fairness_metrics(
  alpha = 0.05
 )
 
-           Fairness Assesment                                  Metric
+           Fairness Assessment                                 Metric
 1          Statistical Parity                Positive Prediction Rate
 2           Equal Opportunity                     False Negative Rate
 3         Predictive Equality                     False Positive Rate
