+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: fairmetrics
 Title: Fairness Evaluation Metrics with Confidence Intervals for Binary Protected Attributes
-Version: 1.0.5
+Version: 1.0.6
 Authors@R: 
     c(
     person("Jianhui", "Gao", email = "jianhui.gao@mail.utoronto.ca", role = c("aut"),comment = c(ORCID = "0000-0003-0915-1473")),
