+++ utils/joss_paper/paper.bib
@@ -172,7 +172,7 @@ series = {ITCS '12}
 @article{Castelnovo_Crupi_Greco_Regoli_Penco_Cosentini_2022, title={A clarification of the nuances in the fairness metrics landscape}, volume={12}, url={https://www.nature.com/articles/s41598-022-07939-1}, DOI={10.1038/s41598-022-07939-1}, number={1}, journal={Scientific Reports}, author={Castelnovo, Alessandro and Crupi, Riccardo and Greco, Greta and Regoli, Daniele and Penco, Ilaria Giuseppina and Cosentini, Andrea Claudio}, year={2022}, month=mar }
 
  @Manual{fairness_package,
-    title = {{f}airness: Algorithmic Fairness Metrics},
+    title = {{fairness}: Algorithmic Fairness Metrics},
     author = {Nikita Kozodoi and Tibor {V. Varga}},
     year = {2021},
     note = {R package version 1.2.1},
@@ -211,7 +211,7 @@ series = {ITCS '12}
 }
 
 @article{wisniewski2022fairmodels,
-  title     = {{f}airmodels: a Flexible Tool for Bias Detection, Visualization, and Mitigation in Binary Classification Models},
+  title     = {{fairmodels}: a Flexible Tool for Bias Detection, Visualization, and Mitigation in Binary Classification Models},
   author    = {Jakub Wiśniewski and Przemysław Biecek},
   journal   = {The R Journal},
   year      = {2022},
@@ -258,7 +258,7 @@ series = {ITCS '12}
 
   @Manual{
   mlr3fairness_package,
-  title = {mlr3fairness: Fairness Auditing and Debiasing for 'mlr3'},
+  title = {{mlr3fairness}: Fairness Auditing and Debiasing for 'mlr3'},
   author = {Florian Pfisterer and Wei Siyi and Michel Lang},
   year = {2024},
   note = {R package version 0.3.2, https://github.com/mlr-org/mlr3fairness},
