+++ utils/joss_paper/paper.tex
@@ -320,7 +320,8 @@ protected attributes. The package provides both point and interval
 estimates for a variety of commonly used criteria. \texttt{fairmetrics}
 also includes an example dataset derived from the Medical Information
 Mart for Intensive Care, version II (MIMIC-II) database (Goldberger et
-al. 2000; J. Raffa 2016) to demonstrate its use.
+al. 2000; J. Raffa 2016; J. D. Raffa et al. 2016) to demonstrate its
+use.
 
 \section{Statement of Need}\label{statement-of-need}
 
@@ -415,7 +416,7 @@ not include one.
  \AttributeTok{alpha =} \FloatTok{0.05}
 \NormalTok{)}
 
-\NormalTok{           Fairness Assesment                                  Metric}
+\NormalTok{           Fairness Assessment                                 Metric}
 \DecValTok{1}\NormalTok{          Statistical Parity                Positive Prediction Rate}
 \DecValTok{2}\NormalTok{           Equal Opportunity                     False Negative Rate}
 \DecValTok{3}\NormalTok{         Predictive Equality                     False Positive Rate}
@@ -512,13 +513,13 @@ Complex Physiologic Signals.''} \emph{Circulation {[}Online{]}} 101
 (23): e215--20. \url{https://doi.org/10.1161/01.CIR.101.23.e215}.
 
 \bibitem[\citeproctext]{ref-fairness_package}
-Kozodoi, Nikita, and Tibor V. Varga. 2021. \emph{Fairness: Algorithmic
+Kozodoi, Nikita, and Tibor V. Varga. 2021. \emph{{fairness}: Algorithmic
 Fairness Metrics}. \url{https://doi.org/10.32614/cran.package.fairness}.
 
 \bibitem[\citeproctext]{ref-mattuMachineBias}
 Mattu, Lauren Kirchner, Jeff Larson. 2016. {``Machine {Bias}.''}
 \emph{ProPublica}.
-https://www.propublica.org/article/machine-bias-risk-assessments-in-criminal-sentencing.
+\url{https://www.propublica.org/article/machine-bias-risk-assessments-in-criminal-sentencing}.
 
 \bibitem[\citeproctext]{ref-mehrabi_survey_21}
 Mehrabi, Ninareh, Fred Morstatter, Nripsuta Saxena, Kristina Lerman, and
@@ -527,14 +528,13 @@ Learning.''} \emph{ACM Comput. Surv.} 54 (6).
 \url{https://doi.org/10.1145/3457607}.
 
 \bibitem[\citeproctext]{ref-mlr3fairness_package}
-Pfisterer, Florian, Wei Siyi, and Michel Lang. 2024. \emph{Mlr3fairness:
-Fairness Auditing and Debiasing for 'Mlr3'}.
+Pfisterer, Florian, Wei Siyi, and Michel Lang. 2024.
+\emph{{mlr3fairness}: Fairness Auditing and Debiasing for 'Mlr3'}.
 \url{https://doi.org/10.32614/cran.package.mlr3fairness}.
 
 \bibitem[\citeproctext]{ref-raffa2016clinical}
 Raffa, Jesse. 2016. {``Clinical Data from the MIMIC-II Database for a
 Case Study on Indwelling Arterial Catheters (Version 1.0).''}
-\url{https://doi.org/10.13026/C2NC7F}.
 \url{https://doi.org/10.13026/C2NC7F}.
 
 \bibitem[\citeproctext]{ref-raffa2016data}
@@ -550,7 +550,7 @@ of AI Systems.''} \emph{arXiv.org}.
 \url{https://doi.org/10.48550/arXiv.2303.16626}.
 
[…]

[…]
