### 2025-09-15 349fb48
using the abbreviation 'CI' for confidence interval
