+++ utils/joss_paper/paper.Rmd
@@ -24,12 +24,12 @@ affiliations:
 date: "2025-06-10"
 bibliography: paper.bib
 output:
+  rticles::joss_article:
+    keep_tex: true
   md_document:
      #keep_tex: true
      preserve_yaml: TRUE
      variant: "markdown_strict"
-  rticles::joss_article:
-    keep_tex: true
 journal: JOSS
 ---
 
