+++ utils/joss_paper/paper.tex
@@ -401,9 +401,6 @@ zero or when the interval for a ratio-based metric does not include one.
 
 \begin{Shaded}
 \begin{Highlighting}[]
-\CommentTok{\# Train a classification model (e.g., random forest).}
-\CommentTok{\# Add the vector of predicted probabilities to the test data}
-\CommentTok{\# to evaluate fairness.}
 \FunctionTok{library}\NormalTok{(fairmetrics)}
 \CommentTok{\# Setting alpha=0.05 for 95\% CIs}
 \FunctionTok{get\_fairness\_metrics}\NormalTok{(}
