+++ utils/joss_paper/paper.md
@@ -24,8 +24,8 @@ affiliations:
 date: "2025-06-10"
 bibliography: paper.bib
 output:
-  rticles::joss_article:
-    keep_tex: true
+  # rticles::joss_article:
+  #   keep_tex: true
   md_document:
      #keep_tex: true
      preserve_yaml: TRUE
