### 2025-09-22 #8497 @benyamindsmith
@Nikoleta-v3 good catch. I believe the issue is because of

```yaml
  rticles::joss_article:
    keep_tex: true
```

I have commented those lines out of the YAML header and I believe the paper should build from there.

For reference. I based my my template off of [this paper](https://github.com/rempsyc/rempsyc/blob/main/paper/paper.md?plain=1)

Let me know if it works.
