### 2025-09-22 e9dda61
removing output from YAML header in md file
