### 2025-09-22 #8497 @benyamindsmith
@Nikoleta-v3 I removed the whole `output:` block from the YAML header in the md file.

Let me know if works now.
