### 2025-09-22 0e4680a
removing \ref latex syntax
