### 2025-09-22 #8497 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
