### 2025-09-22 0c7ebdd
Adding JOSS badge to README
